Accuracy at Each Cutoff.

Find the accuracy at each possible cutoff. Actuals should be binary, where 1 = present and 0 = absent.

Usage

accuracyAtEachCutoff(
  predicted,
  actual,
  UH = NULL,
  UM = NULL,
  UCR = NULL,
  UFA = NULL
)

Arguments

predicted

vector of continuous predicted values.

actual

vector of binary actual values (1 = present and 0 = absent).

UH

(optional) utility of hits (true positives), specified as a value from 0-1, where 1 is the most highly valued and 0 is the least valued.

UM

(optional) utility of misses (false negatives), specified as a value from 0-1, where 1 is the most highly valued and 0 is the least valued.

UCR

(optional) utility of correct rejections (true negatives), specified as a value from 0-1, where 1 is the most highly valued and 0 is the least valued.

UFA

(optional) utility of false positives (false positives), specified as a value from 0-1, where 1 is the most highly valued and 0 is the least valued.

Value

cutoff = the cutoff specified

TP = true positives

TN = true negatives

FP = false positives

FN = false negatives

SR = selection ratio

BR = base rate

percentAccuracy = percent accuracy

percentAccuracyByChance = percent accuracy by chance

percentAccuracyPredictingFromBaseRate = percent accuracy from predicting from the base rate

RIOC = relative improvement over chance

relativeImprovementOverPredictingFromBaseRate = relative improvement over predicting from the base rate

SN = sensitivty

SP = specificity

TPrate = true positive rate

TNrate = true negative rate

FNrate = false negative rate

FPrate = false positive rate

HR = hit rate

FAR = false alarm rate

PPV = positive predictive value

NPV = negative predictive value

FDR = false discovery rate

FOR = false omission rate

youdenJ = Youden's J statistic

balancedAccuracy = balanced accuracy

f1Score = F1-score

mcc = Matthews correlation coefficient

diagnosticOddsRatio = diagnostic odds ratio

positiveLikelihoodRatio = positive likelihood ratio

negativeLikelhoodRatio = negative likelihood ratio

dPrimeSDT = d-Prime index from signal detection theory

betaSDT = beta index from signal detection theory

cSDT = c index from signal detection theory

aSDT = a index from signal detection theory

bSDT = b index from signal detection theory

differenceBetweenPredictedAndObserved = difference between predicted and observed values

informationGain = information gain

overallUtility = overall utility (if utilities were specified)

Details

Compute accuracy indices of predicted values in relation to actual values at each possible cutoff by specifying the predicted values and actual values. The target condition is considered present at or above each cutoff value. Optionally, you can specify the utility of hits, misses, correct rejections, and false alarms to calculate the overall utility of each possible cutoff.

See also

Examples

# Prepare Data
data("USArrests")
USArrests$highMurderState <- NA
USArrests$highMurderState[which(USArrests$Murder >= 10)] <- 1
USArrests$highMurderState[which(USArrests$Murder < 10)] <- 0

# Calculate Accuracy
accuracyAtEachCutoff(predicted = USArrests$Assault,
  actual = USArrests$highMurderState)
#>    cutoff TP TN FP FN   SR   BR percentAccuracy percentAccuracyByChance
#> 1   45.00 16  0 34  0 1.00 0.32              32                   32.00
#> 2   46.00 16  1 33  0 0.98 0.32              34                   32.72
#> 3   48.00 16  2 32  0 0.96 0.32              36                   33.44
#> 4   53.00 16  3 31  0 0.94 0.32              38                   34.16
#> 5   56.00 16  4 30  0 0.92 0.32              40                   34.88
#> 6   57.00 16  5 29  0 0.90 0.32              42                   35.60
#> 7   72.00 16  6 28  0 0.88 0.32              44                   36.32
#> 8   81.00 16  7 27  0 0.86 0.32              46                   37.04
#> 9   83.00 16  8 26  0 0.84 0.32              48                   37.76
#> 10  86.00 16  9 25  0 0.82 0.32              50                   38.48
#> 11 102.00 16 10 24  0 0.80 0.32              52                   39.20
#> 12 106.00 16 11 23  0 0.78 0.32              54                   39.92
#> 13 109.00 16 12 22  0 0.76 0.32              56                   40.64
#> 14 110.00 16 14 20  0 0.72 0.32              60                   42.08
#> 15 113.00 16 15 19  0 0.70 0.32              62                   42.80
#> 16 115.00 16 16 18  0 0.68 0.32              64                   43.52
#> 17 120.00 16 17 17  0 0.66 0.32              66                   44.24
#> 18 145.00 16 20 14  0 0.60 0.32              72                   46.40
#> 19 149.00 16 21 13  0 0.58 0.32              74                   47.12
#> 20 151.00 16 22 12  0 0.56 0.32              76                   47.84
#> 21 156.00 16 23 11  0 0.54 0.32              78                   48.56
#> 22 159.00 16 24 10  0 0.52 0.32              80                   49.28
#> 23 161.00 16 26  8  0 0.48 0.32              84                   50.72
#> 24 174.00 16 27  7  0 0.46 0.32              86                   51.44
#> 25 178.00 16 28  6  0 0.44 0.32              88                   52.16
#> 26 188.00 16 29  5  0 0.42 0.32              90                   52.88
#> 27 190.00 15 29  5  1 0.40 0.32              88                   53.60
#> 28 201.00 15 30  4  1 0.38 0.32              90                   54.32
#> 29 204.00 14 30  4  2 0.36 0.32              88                   55.04
#> 30 211.00 14 31  3  2 0.34 0.32              90                   55.76
#> 31 236.00 13 31  3  3 0.32 0.32              88                   56.48
#> 32 238.00 12 31  3  4 0.30 0.32              86                   57.20
#> 33 249.00 12 32  2  4 0.28 0.32              88                   57.92
#> 34 252.00 10 32  2  6 0.24 0.32              84                   59.36
#> 35 254.00  9 32  2  7 0.22 0.32              82                   60.08
#> 36 255.00  8 32  2  8 0.20 0.32              80                   60.80
#> 37 259.00  7 32  2  9 0.18 0.32              78                   61.52
#> 38 263.00  6 32  2 10 0.16 0.32              76                   62.24
#> 39 276.00  5 32  2 11 0.14 0.32              74                   62.96
#> 40 279.00  5 33  1 11 0.12 0.32              76                   63.68
#> 41 285.00  4 33  1 12 0.10 0.32              74                   64.40
#> 42 294.00  3 33  1 13 0.08 0.32              72                   65.12
#> 43 300.00  3 34  0 13 0.06 0.32              74                   65.84
#> 44 335.00  2 34  0 14 0.04 0.32              72                   66.56
#> 45 337.00  1 34  0 15 0.02 0.32              70                   67.28
#> 46 337.01  0 34  0 16 0.00 0.32              68                   68.00
#>    percentAccuracyPredictingFromBaseRate       RIOC
#> 1                                     68         NA
#> 2                                     68 1.00000000
#> 3                                     68 1.00000000
#> 4                                     68 1.00000000
#> 5                                     68 1.00000000
#> 6                                     68 1.00000000
#> 7                                     68 1.00000000
#> 8                                     68 1.00000000
#> 9                                     68 1.00000000
#> 10                                    68 1.00000000
#> 11                                    68 1.00000000
#> 12                                    68 1.00000000
#> 13                                    68 1.00000000
#> 14                                    68 1.00000000
#> 15                                    68 1.00000000
#> 16                                    68 1.00000000
#> 17                                    68 1.00000000
#> 18                                    68 1.00000000
#> 19                                    68 1.00000000
#> 20                                    68 1.00000000
#> 21                                    68 1.00000000
#> 22                                    68 1.00000000
#> 23                                    68 1.00000000
#> 24                                    68 1.00000000
#> 25                                    68 1.00000000
#> 26                                    68 1.00000000
#> 27                                    68 0.89583333
#> 28                                    68 0.89919355
#> 29                                    68 0.80468750
#> 30                                    68 0.81060606
#> 31                                    68 0.72426471
#> 32                                    68 0.64285714
#> 33                                    68 0.65277778
#> 34                                    68 0.50657895
#> 35                                    68 0.43910256
#> 36                                    68 0.37500000
#> 37                                    68 0.31402439
#> 38                                    68 0.25595238
#> 39                                    68 0.20058140
#> 40                                    68 0.21875000
#> 41                                    68 0.16666667
#> 42                                    68 0.11684783
#> 43                                    68 0.13563830
#> 44                                    68 0.08854167
#> 45                                    68 0.04336735
#> 46                                    68 0.00000000
#>    relativeImprovementOverPredictingFromBaseRate     SN         SP TPrate
#> 1                                       -0.56250 1.0000 0.00000000 1.0000
#> 2                                       -0.53125 1.0000 0.02941176 1.0000
#> 3                                       -0.50000 1.0000 0.05882353 1.0000
#> 4                                       -0.46875 1.0000 0.08823529 1.0000
#> 5                                       -0.43750 1.0000 0.11764706 1.0000
#> 6                                       -0.40625 1.0000 0.14705882 1.0000
#> 7                                       -0.37500 1.0000 0.17647059 1.0000
#> 8                                       -0.34375 1.0000 0.20588235 1.0000
#> 9                                       -0.31250 1.0000 0.23529412 1.0000
#> 10                                      -0.28125 1.0000 0.26470588 1.0000
#> 11                                      -0.25000 1.0000 0.29411765 1.0000
#> 12                                      -0.21875 1.0000 0.32352941 1.0000
#> 13                                      -0.18750 1.0000 0.35294118 1.0000
#> 14                                      -0.12500 1.0000 0.41176471 1.0000
#> 15                                      -0.09375 1.0000 0.44117647 1.0000
#> 16                                      -0.06250 1.0000 0.47058824 1.0000
#> 17                                      -0.03125 1.0000 0.50000000 1.0000
#> 18                                       0.06250 1.0000 0.58823529 1.0000
#> 19                                       0.09375 1.0000 0.61764706 1.0000
#> 20                                       0.12500 1.0000 0.64705882 1.0000
#> 21                                       0.15625 1.0000 0.67647059 1.0000
#> 22                                       0.18750 1.0000 0.70588235 1.0000
#> 23                                       0.25000 1.0000 0.76470588 1.0000
#> 24                                       0.28125 1.0000 0.79411765 1.0000
#> 25                                       0.31250 1.0000 0.82352941 1.0000
#> 26                                       0.34375 1.0000 0.85294118 1.0000
#> 27                                       0.31250 0.9375 0.85294118 0.9375
#> 28                                       0.34375 0.9375 0.88235294 0.9375
#> 29                                       0.31250 0.8750 0.88235294 0.8750
#> 30                                       0.34375 0.8750 0.91176471 0.8750
#> 31                                       0.31250 0.8125 0.91176471 0.8125
#> 32                                       0.28125 0.7500 0.91176471 0.7500
#> 33                                       0.31250 0.7500 0.94117647 0.7500
#> 34                                       0.25000 0.6250 0.94117647 0.6250
#> 35                                       0.21875 0.5625 0.94117647 0.5625
#> 36                                       0.18750 0.5000 0.94117647 0.5000
#> 37                                       0.15625 0.4375 0.94117647 0.4375
#> 38                                       0.12500 0.3750 0.94117647 0.3750
#> 39                                       0.09375 0.3125 0.94117647 0.3125
#> 40                                       0.12500 0.3125 0.97058824 0.3125
#> 41                                       0.09375 0.2500 0.97058824 0.2500
#> 42                                       0.06250 0.1875 0.97058824 0.1875
#> 43                                       0.09375 0.1875 1.00000000 0.1875
#> 44                                       0.06250 0.1250 1.00000000 0.1250
#> 45                                       0.03125 0.0625 1.00000000 0.0625
#> 46                                       0.00000 0.0000 1.00000000 0.0000
#>        TNrate FNrate     FPrate     HR        FAR       PPV       NPV       FDR
#> 1  0.00000000 0.0000 1.00000000 1.0000 1.00000000 0.3200000        NA 0.6800000
#> 2  0.02941176 0.0000 0.97058824 1.0000 0.97058824 0.3265306 1.0000000 0.6734694
#> 3  0.05882353 0.0000 0.94117647 1.0000 0.94117647 0.3333333 1.0000000 0.6666667
#> 4  0.08823529 0.0000 0.91176471 1.0000 0.91176471 0.3404255 1.0000000 0.6595745
#> 5  0.11764706 0.0000 0.88235294 1.0000 0.88235294 0.3478261 1.0000000 0.6521739
#> 6  0.14705882 0.0000 0.85294118 1.0000 0.85294118 0.3555556 1.0000000 0.6444444
#> 7  0.17647059 0.0000 0.82352941 1.0000 0.82352941 0.3636364 1.0000000 0.6363636
#> 8  0.20588235 0.0000 0.79411765 1.0000 0.79411765 0.3720930 1.0000000 0.6279070
#> 9  0.23529412 0.0000 0.76470588 1.0000 0.76470588 0.3809524 1.0000000 0.6190476
#> 10 0.26470588 0.0000 0.73529412 1.0000 0.73529412 0.3902439 1.0000000 0.6097561
#> 11 0.29411765 0.0000 0.70588235 1.0000 0.70588235 0.4000000 1.0000000 0.6000000
#> 12 0.32352941 0.0000 0.67647059 1.0000 0.67647059 0.4102564 1.0000000 0.5897436
#> 13 0.35294118 0.0000 0.64705882 1.0000 0.64705882 0.4210526 1.0000000 0.5789474
#> 14 0.41176471 0.0000 0.58823529 1.0000 0.58823529 0.4444444 1.0000000 0.5555556
#> 15 0.44117647 0.0000 0.55882353 1.0000 0.55882353 0.4571429 1.0000000 0.5428571
#> 16 0.47058824 0.0000 0.52941176 1.0000 0.52941176 0.4705882 1.0000000 0.5294118
#> 17 0.50000000 0.0000 0.50000000 1.0000 0.50000000 0.4848485 1.0000000 0.5151515
#> 18 0.58823529 0.0000 0.41176471 1.0000 0.41176471 0.5333333 1.0000000 0.4666667
#> 19 0.61764706 0.0000 0.38235294 1.0000 0.38235294 0.5517241 1.0000000 0.4482759
#> 20 0.64705882 0.0000 0.35294118 1.0000 0.35294118 0.5714286 1.0000000 0.4285714
#> 21 0.67647059 0.0000 0.32352941 1.0000 0.32352941 0.5925926 1.0000000 0.4074074
#> 22 0.70588235 0.0000 0.29411765 1.0000 0.29411765 0.6153846 1.0000000 0.3846154
#> 23 0.76470588 0.0000 0.23529412 1.0000 0.23529412 0.6666667 1.0000000 0.3333333
#> 24 0.79411765 0.0000 0.20588235 1.0000 0.20588235 0.6956522 1.0000000 0.3043478
#> 25 0.82352941 0.0000 0.17647059 1.0000 0.17647059 0.7272727 1.0000000 0.2727273
#> 26 0.85294118 0.0000 0.14705882 1.0000 0.14705882 0.7619048 1.0000000 0.2380952
#> 27 0.85294118 0.0625 0.14705882 0.9375 0.14705882 0.7500000 0.9666667 0.2500000
#> 28 0.88235294 0.0625 0.11764706 0.9375 0.11764706 0.7894737 0.9677419 0.2105263
#> 29 0.88235294 0.1250 0.11764706 0.8750 0.11764706 0.7777778 0.9375000 0.2222222
#> 30 0.91176471 0.1250 0.08823529 0.8750 0.08823529 0.8235294 0.9393939 0.1764706
#> 31 0.91176471 0.1875 0.08823529 0.8125 0.08823529 0.8125000 0.9117647 0.1875000
#> 32 0.91176471 0.2500 0.08823529 0.7500 0.08823529 0.8000000 0.8857143 0.2000000
#> 33 0.94117647 0.2500 0.05882353 0.7500 0.05882353 0.8571429 0.8888889 0.1428571
#> 34 0.94117647 0.3750 0.05882353 0.6250 0.05882353 0.8333333 0.8421053 0.1666667
#> 35 0.94117647 0.4375 0.05882353 0.5625 0.05882353 0.8181818 0.8205128 0.1818182
#> 36 0.94117647 0.5000 0.05882353 0.5000 0.05882353 0.8000000 0.8000000 0.2000000
#> 37 0.94117647 0.5625 0.05882353 0.4375 0.05882353 0.7777778 0.7804878 0.2222222
#> 38 0.94117647 0.6250 0.05882353 0.3750 0.05882353 0.7500000 0.7619048 0.2500000
#> 39 0.94117647 0.6875 0.05882353 0.3125 0.05882353 0.7142857 0.7441860 0.2857143
#> 40 0.97058824 0.6875 0.02941176 0.3125 0.02941176 0.8333333 0.7500000 0.1666667
#> 41 0.97058824 0.7500 0.02941176 0.2500 0.02941176 0.8000000 0.7333333 0.2000000
#> 42 0.97058824 0.8125 0.02941176 0.1875 0.02941176 0.7500000 0.7173913 0.2500000
#> 43 1.00000000 0.8125 0.00000000 0.1875 0.00000000 1.0000000 0.7234043 0.0000000
#> 44 1.00000000 0.8750 0.00000000 0.1250 0.00000000 1.0000000 0.7083333 0.0000000
#> 45 1.00000000 0.9375 0.00000000 0.0625 0.00000000 1.0000000 0.6938776 0.0000000
#> 46 1.00000000 1.0000 0.00000000 0.0000 0.00000000        NA 0.6800000        NA
#>           FOR    youdenJ balancedAccuracy   f1Score        mcc
#> 1          NA 0.00000000        0.5000000 0.4848485         NA
#> 2  0.00000000 0.02941176        0.5147059 0.4923077 0.09799919
#> 3  0.00000000 0.05882353        0.5294118 0.5000000 0.14002801
#> 4  0.00000000 0.08823529        0.5441176 0.5079365 0.17331344
#> 5  0.00000000 0.11764706        0.5588235 0.5161290 0.20228869
#> 6  0.00000000 0.14705882        0.5735294 0.5245902 0.22866478
#> 7  0.00000000 0.17647059        0.5882353 0.5333333 0.25332020
#> 8  0.00000000 0.20588235        0.6029412 0.5423729 0.27678040
#> 9  0.00000000 0.23529412        0.6176471 0.5517241 0.29939248
#> 10 0.00000000 0.26470588        0.6323529 0.5614035 0.32140295
#> 11 0.00000000 0.29411765        0.6470588 0.5714286 0.34299717
#> 12 0.00000000 0.32352941        0.6617647 0.5818182 0.36432131
#> 13 0.00000000 0.35294118        0.6764706 0.5925926 0.38549554
#> 14 0.00000000 0.41176471        0.7058824 0.6153846 0.42779263
#> 15 0.00000000 0.44117647        0.7205882 0.6274510 0.44908871
#> 16 0.00000000 0.47058824        0.7352941 0.6400000 0.47058824
#> 17 0.00000000 0.50000000        0.7500000 0.6530612 0.49236596
#> 18 0.00000000 0.58823529        0.7941176 0.6956522 0.56011203
#> 19 0.00000000 0.61764706        0.8088235 0.7111111 0.58375576
#> 20 0.00000000 0.64705882        0.8235294 0.7272727 0.60806899
#> 21 0.00000000 0.67647059        0.8382353 0.7441860 0.63314411
#> 22 0.00000000 0.70588235        0.8529412 0.7619048 0.65908204
#> 23 0.00000000 0.76470588        0.8823529 0.8000000 0.71400555
#> 24 0.00000000 0.79411765        0.8970588 0.8205128 0.74325613
#> 25 0.00000000 0.82352941        0.9117647 0.8421053 0.77390599
#> 26 0.00000000 0.85294118        0.9264706 0.8648649 0.80613891
#> 27 0.03333333 0.79044118        0.8952206 0.8333333 0.75265055
#> 28 0.03225806 0.81985294        0.9099265 0.8571429 0.78791208
#> 29 0.06250000 0.75735294        0.8786765 0.8235294 0.73601476
#> 30 0.06060606 0.78676471        0.8933824 0.8484848 0.77475233
#> 31 0.08823529 0.72426471        0.8621324 0.8125000 0.72426471
#> 32 0.11428571 0.66176471        0.8308824 0.7741935 0.67363307
#> 33 0.11111111 0.69117647        0.8455882 0.8000000 0.71808049
#> 34 0.15789474 0.56617647        0.7830882 0.7142857 0.61839909
#> 35 0.17948718 0.50367647        0.7518382 0.6666667 0.56718204
#> 36 0.20000000 0.44117647        0.7205882 0.6153846 0.51449576
#> 37 0.21951220 0.37867647        0.6893382 0.5600000 0.45978478
#> 38 0.23809524 0.31617647        0.6580882 0.5000000 0.40230864
#> 39 0.25581395 0.25367647        0.6268382 0.4347826 0.34103299
#> 40 0.25000000 0.28308824        0.6415441 0.4545455 0.40636782
#> 41 0.26666667 0.22058824        0.6102941 0.3809524 0.34299717
#> 42 0.28260870 0.15808824        0.5790441 0.3000000 0.27182543
#> 43 0.27659574 0.18750000        0.5937500 0.3157895 0.36829105
#> 44 0.29166667 0.12500000        0.5625000 0.2222222 0.29755952
#> 45 0.30612245 0.06250000        0.5312500 0.1176471 0.20824828
#> 46 0.32000000 0.00000000        0.5000000        NA         NA
#>    diagnosticOddsRatio positiveLikelihoodRatio negativeLikelihoodRatio
#> 1                   NA                1.000000                      NA
#> 2                   NA                1.030303              0.00000000
#> 3                   NA                1.062500              0.00000000
#> 4                   NA                1.096774              0.00000000
#> 5                   NA                1.133333              0.00000000
#> 6                   NA                1.172414              0.00000000
#> 7                   NA                1.214286              0.00000000
#> 8                   NA                1.259259              0.00000000
#> 9                   NA                1.307692              0.00000000
#> 10                  NA                1.360000              0.00000000
#> 11                  NA                1.416667              0.00000000
#> 12                  NA                1.478261              0.00000000
#> 13                  NA                1.545455              0.00000000
#> 14                  NA                1.700000              0.00000000
#> 15                  NA                1.789474              0.00000000
#> 16                  NA                1.888889              0.00000000
#> 17                  NA                2.000000              0.00000000
#> 18                  NA                2.428571              0.00000000
#> 19                  NA                2.615385              0.00000000
#> 20                  NA                2.833333              0.00000000
#> 21                  NA                3.090909              0.00000000
#> 22                  NA                3.400000              0.00000000
#> 23                  NA                4.250000              0.00000000
#> 24                  NA                4.857143              0.00000000
#> 25                  NA                5.666667              0.00000000
#> 26                  NA                6.800000              0.00000000
#> 27           87.000000                6.375000              0.07327586
#> 28          112.500000                7.968750              0.07083333
#> 29           52.500000                7.437500              0.14166667
#> 30           72.333333                9.916667              0.13709677
#> 31           44.777778                9.208333              0.20564516
#> 32           31.000000                8.500000              0.27419355
#> 33           48.000000               12.750000              0.26562500
#> 34           26.666667               10.625000              0.39843750
#> 35           20.571429                9.562500              0.46484375
#> 36           16.000000                8.500000              0.53125000
#> 37           12.444444                7.437500              0.59765625
#> 38            9.600000                6.375000              0.66406250
#> 39            7.272727                5.312500              0.73046875
#> 40           15.000000               10.625000              0.70833333
#> 41           11.000000                8.500000              0.77272727
#> 42            7.615385                6.375000              0.83712121
#> 43                  NA                      NA              0.81250000
#> 44                  NA                      NA              0.87500000
#> 45                  NA                      NA              0.93750000
#> 46                  NA                      NA              1.00000000
#>    dPrimeSDT   betaSDT        cSDT      aSDT        bSDT
#> 1         NA        NA          NA        NA          NA
#> 2         NA 0.0000000          NA 0.7573529  0.02857143
#> 3         NA 0.0000000          NA 0.7647059  0.05555556
#> 4         NA 0.0000000          NA 0.7720588  0.08108108
#> 5         NA 0.0000000          NA 0.7794118  0.10526316
#> 6         NA 0.0000000          NA 0.7867647  0.12820513
#> 7         NA 0.0000000          NA 0.7941176  0.15000000
#> 8         NA 0.0000000          NA 0.8014706  0.17073171
#> 9         NA 0.0000000          NA 0.8088235  0.19047619
#> 10        NA 0.0000000          NA 0.8161765  0.20930233
#> 11        NA 0.0000000          NA 0.8235294  0.22727273
#> 12        NA 0.0000000          NA 0.8308824  0.24444444
#> 13        NA 0.0000000          NA 0.8382353  0.26086957
#> 14        NA 0.0000000          NA 0.8529412  0.29166667
#> 15        NA 0.0000000          NA 0.8602941  0.30612245
#> 16        NA 0.0000000          NA 0.8676471  0.32000000
#> 17        NA 0.0000000          NA 0.8750000  0.33333333
#> 18        NA 0.0000000          NA 0.8970588  0.37777778
#> 19        NA 0.0000000          NA 0.9044118  0.39534884
#> 20        NA 0.0000000          NA 0.9117647  0.41463415
#> 21        NA 0.0000000          NA 0.9191176  0.43589744
#> 22        NA 0.0000000          NA 0.9264706  0.45945946
#> 23        NA 0.0000000          NA 0.9411765  0.51515152
#> 24        NA 0.0000000          NA 0.9485294  0.54838710
#> 25        NA 0.0000000          NA 0.9558824  0.58620690
#> 26        NA 0.0000000          NA 0.9632353  0.62962963
#> 27  2.583252 0.5344995 -0.24249457 0.9384191  0.78703704
#> 28  2.720952 0.6234551 -0.17364456 0.9476103  0.85000000
#> 29  2.337181 1.0435544  0.01824103 0.9246324  1.02000000
#> 30  2.502052 1.2864643  0.10067643 0.9356618  1.10869565
#> 31  2.238849 1.6820865  0.23227784 0.9145221  1.29347826
#> 32  2.026192 1.9859180  0.33860624 0.8933824  1.47826087
#> 33  2.239216 2.7093703  0.44511836 0.9080882  1.61904762
#> 34  1.883366 3.2330273  0.62304355 0.8694853  2.02380952
#> 35  1.722037 3.3595638  0.70370789 0.8501838  2.22619048
#> 36  1.564726 3.4013910  0.78236324 0.8308824  2.42857143
#> 37  1.407416 3.3595638  0.86101858 0.8110557  2.51331497
#> 38  1.246087 3.2330273  0.94168292 0.7898284  2.58525346
#> 39  1.075950 3.0184193  1.02675144 0.7663603  2.62114537
#> 40  1.400734 5.2892304  1.18914319 0.7972426  3.22784810
#> 41  1.215020 4.7476784  1.28199986 0.7757353  3.40000000
#> 42  1.002363 4.0213169  1.38832826 0.7503064  3.44839858
#> 43        NA        NA          NA 0.7968750  6.33333333
#> 44        NA        NA          NA 0.7812500  9.00000000
#> 45        NA        NA          NA 0.7656250 17.00000000
#> 46        NA        NA          NA        NA          NA
#>    differenceBetweenPredictedAndObserved informationGain
#> 1                                49.6000              NA
#> 2                                49.6000              NA
#> 3                                49.6000              NA
#> 4                                49.6000              NA
#> 5                                49.6000              NA
#> 6                                75.8000              NA
#> 7                                75.8000              NA
#> 8                                75.8000              NA
#> 9                                75.8000              NA
#> 10                               75.8000              NA
#> 11                              107.2000              NA
#> 12                              107.2000              NA
#> 13                              107.2000              NA
#> 14                              107.2000              NA
#> 15                              117.6000              NA
#> 16                              117.6000              NA
#> 17                              117.6000              NA
#> 18                              153.1667              NA
#> 19                              153.1667              NA
#> 20                              153.1667              NA
#> 21                              153.1667              NA
#> 22                              175.0000              NA
#> 23                              175.0000              NA
#> 24                              175.0000              NA
#> 25                              175.0000              NA
#> 26                              175.0000              NA
#> 27                              207.8000      0.45336483
#> 28                              207.8000      0.49476883
#> 29                              207.8000      0.41340219
#> 30                              207.8000      0.45810281
#> 31                              207.8000      0.38881826
#> 32                              247.6000      0.32890663
#> 33                              247.6000      0.37636708
#> 34                              247.6000      0.27014667
#> 35                              265.6000      0.22430896
#> 36                              265.6000      0.18245336
#> 37                              265.6000      0.14421850
#> 38                              265.6000      0.10941594
#> 39                              265.6000      0.07803185
#> 40                              265.6000      0.11245402
#> 41                              309.4000      0.07921198
#> 42                              309.4000      0.04921665
#> 43                              309.4000              NA
#> 44                              309.4000              NA
#> 45                              309.4000              NA
#> 46                                    NA              NA
accuracyAtEachCutoff(predicted = USArrests$Assault,
  actual = USArrests$highMurderState,
  UH = 1, UM = 0, UCR = .9, UFA = 0)
#>    cutoff TP TN FP FN   SR   BR percentAccuracy percentAccuracyByChance
#> 1   45.00 16  0 34  0 1.00 0.32              32                   32.00
#> 2   46.00 16  1 33  0 0.98 0.32              34                   32.72
#> 3   48.00 16  2 32  0 0.96 0.32              36                   33.44
#> 4   53.00 16  3 31  0 0.94 0.32              38                   34.16
#> 5   56.00 16  4 30  0 0.92 0.32              40                   34.88
#> 6   57.00 16  5 29  0 0.90 0.32              42                   35.60
#> 7   72.00 16  6 28  0 0.88 0.32              44                   36.32
#> 8   81.00 16  7 27  0 0.86 0.32              46                   37.04
#> 9   83.00 16  8 26  0 0.84 0.32              48                   37.76
#> 10  86.00 16  9 25  0 0.82 0.32              50                   38.48
#> 11 102.00 16 10 24  0 0.80 0.32              52                   39.20
#> 12 106.00 16 11 23  0 0.78 0.32              54                   39.92
#> 13 109.00 16 12 22  0 0.76 0.32              56                   40.64
#> 14 110.00 16 14 20  0 0.72 0.32              60                   42.08
#> 15 113.00 16 15 19  0 0.70 0.32              62                   42.80
#> 16 115.00 16 16 18  0 0.68 0.32              64                   43.52
#> 17 120.00 16 17 17  0 0.66 0.32              66                   44.24
#> 18 145.00 16 20 14  0 0.60 0.32              72                   46.40
#> 19 149.00 16 21 13  0 0.58 0.32              74                   47.12
#> 20 151.00 16 22 12  0 0.56 0.32              76                   47.84
#> 21 156.00 16 23 11  0 0.54 0.32              78                   48.56
#> 22 159.00 16 24 10  0 0.52 0.32              80                   49.28
#> 23 161.00 16 26  8  0 0.48 0.32              84                   50.72
#> 24 174.00 16 27  7  0 0.46 0.32              86                   51.44
#> 25 178.00 16 28  6  0 0.44 0.32              88                   52.16
#> 26 188.00 16 29  5  0 0.42 0.32              90                   52.88
#> 27 190.00 15 29  5  1 0.40 0.32              88                   53.60
#> 28 201.00 15 30  4  1 0.38 0.32              90                   54.32
#> 29 204.00 14 30  4  2 0.36 0.32              88                   55.04
#> 30 211.00 14 31  3  2 0.34 0.32              90                   55.76
#> 31 236.00 13 31  3  3 0.32 0.32              88                   56.48
#> 32 238.00 12 31  3  4 0.30 0.32              86                   57.20
#> 33 249.00 12 32  2  4 0.28 0.32              88                   57.92
#> 34 252.00 10 32  2  6 0.24 0.32              84                   59.36
#> 35 254.00  9 32  2  7 0.22 0.32              82                   60.08
#> 36 255.00  8 32  2  8 0.20 0.32              80                   60.80
#> 37 259.00  7 32  2  9 0.18 0.32              78                   61.52
#> 38 263.00  6 32  2 10 0.16 0.32              76                   62.24
#> 39 276.00  5 32  2 11 0.14 0.32              74                   62.96
#> 40 279.00  5 33  1 11 0.12 0.32              76                   63.68
#> 41 285.00  4 33  1 12 0.10 0.32              74                   64.40
#> 42 294.00  3 33  1 13 0.08 0.32              72                   65.12
#> 43 300.00  3 34  0 13 0.06 0.32              74                   65.84
#> 44 335.00  2 34  0 14 0.04 0.32              72                   66.56
#> 45 337.00  1 34  0 15 0.02 0.32              70                   67.28
#> 46 337.01  0 34  0 16 0.00 0.32              68                   68.00
#>    percentAccuracyPredictingFromBaseRate       RIOC
#> 1                                     68         NA
#> 2                                     68 1.00000000
#> 3                                     68 1.00000000
#> 4                                     68 1.00000000
#> 5                                     68 1.00000000
#> 6                                     68 1.00000000
#> 7                                     68 1.00000000
#> 8                                     68 1.00000000
#> 9                                     68 1.00000000
#> 10                                    68 1.00000000
#> 11                                    68 1.00000000
#> 12                                    68 1.00000000
#> 13                                    68 1.00000000
#> 14                                    68 1.00000000
#> 15                                    68 1.00000000
#> 16                                    68 1.00000000
#> 17                                    68 1.00000000
#> 18                                    68 1.00000000
#> 19                                    68 1.00000000
#> 20                                    68 1.00000000
#> 21                                    68 1.00000000
#> 22                                    68 1.00000000
#> 23                                    68 1.00000000
#> 24                                    68 1.00000000
#> 25                                    68 1.00000000
#> 26                                    68 1.00000000
#> 27                                    68 0.89583333
#> 28                                    68 0.89919355
#> 29                                    68 0.80468750
#> 30                                    68 0.81060606
#> 31                                    68 0.72426471
#> 32                                    68 0.64285714
#> 33                                    68 0.65277778
#> 34                                    68 0.50657895
#> 35                                    68 0.43910256
#> 36                                    68 0.37500000
#> 37                                    68 0.31402439
#> 38                                    68 0.25595238
#> 39                                    68 0.20058140
#> 40                                    68 0.21875000
#> 41                                    68 0.16666667
#> 42                                    68 0.11684783
#> 43                                    68 0.13563830
#> 44                                    68 0.08854167
#> 45                                    68 0.04336735
#> 46                                    68 0.00000000
#>    relativeImprovementOverPredictingFromBaseRate     SN         SP TPrate
#> 1                                       -0.56250 1.0000 0.00000000 1.0000
#> 2                                       -0.53125 1.0000 0.02941176 1.0000
#> 3                                       -0.50000 1.0000 0.05882353 1.0000
#> 4                                       -0.46875 1.0000 0.08823529 1.0000
#> 5                                       -0.43750 1.0000 0.11764706 1.0000
#> 6                                       -0.40625 1.0000 0.14705882 1.0000
#> 7                                       -0.37500 1.0000 0.17647059 1.0000
#> 8                                       -0.34375 1.0000 0.20588235 1.0000
#> 9                                       -0.31250 1.0000 0.23529412 1.0000
#> 10                                      -0.28125 1.0000 0.26470588 1.0000
#> 11                                      -0.25000 1.0000 0.29411765 1.0000
#> 12                                      -0.21875 1.0000 0.32352941 1.0000
#> 13                                      -0.18750 1.0000 0.35294118 1.0000
#> 14                                      -0.12500 1.0000 0.41176471 1.0000
#> 15                                      -0.09375 1.0000 0.44117647 1.0000
#> 16                                      -0.06250 1.0000 0.47058824 1.0000
#> 17                                      -0.03125 1.0000 0.50000000 1.0000
#> 18                                       0.06250 1.0000 0.58823529 1.0000
#> 19                                       0.09375 1.0000 0.61764706 1.0000
#> 20                                       0.12500 1.0000 0.64705882 1.0000
#> 21                                       0.15625 1.0000 0.67647059 1.0000
#> 22                                       0.18750 1.0000 0.70588235 1.0000
#> 23                                       0.25000 1.0000 0.76470588 1.0000
#> 24                                       0.28125 1.0000 0.79411765 1.0000
#> 25                                       0.31250 1.0000 0.82352941 1.0000
#> 26                                       0.34375 1.0000 0.85294118 1.0000
#> 27                                       0.31250 0.9375 0.85294118 0.9375
#> 28                                       0.34375 0.9375 0.88235294 0.9375
#> 29                                       0.31250 0.8750 0.88235294 0.8750
#> 30                                       0.34375 0.8750 0.91176471 0.8750
#> 31                                       0.31250 0.8125 0.91176471 0.8125
#> 32                                       0.28125 0.7500 0.91176471 0.7500
#> 33                                       0.31250 0.7500 0.94117647 0.7500
#> 34                                       0.25000 0.6250 0.94117647 0.6250
#> 35                                       0.21875 0.5625 0.94117647 0.5625
#> 36                                       0.18750 0.5000 0.94117647 0.5000
#> 37                                       0.15625 0.4375 0.94117647 0.4375
#> 38                                       0.12500 0.3750 0.94117647 0.3750
#> 39                                       0.09375 0.3125 0.94117647 0.3125
#> 40                                       0.12500 0.3125 0.97058824 0.3125
#> 41                                       0.09375 0.2500 0.97058824 0.2500
#> 42                                       0.06250 0.1875 0.97058824 0.1875
#> 43                                       0.09375 0.1875 1.00000000 0.1875
#> 44                                       0.06250 0.1250 1.00000000 0.1250
#> 45                                       0.03125 0.0625 1.00000000 0.0625
#> 46                                       0.00000 0.0000 1.00000000 0.0000
#>        TNrate FNrate     FPrate     HR        FAR       PPV       NPV       FDR
#> 1  0.00000000 0.0000 1.00000000 1.0000 1.00000000 0.3200000        NA 0.6800000
#> 2  0.02941176 0.0000 0.97058824 1.0000 0.97058824 0.3265306 1.0000000 0.6734694
#> 3  0.05882353 0.0000 0.94117647 1.0000 0.94117647 0.3333333 1.0000000 0.6666667
#> 4  0.08823529 0.0000 0.91176471 1.0000 0.91176471 0.3404255 1.0000000 0.6595745
#> 5  0.11764706 0.0000 0.88235294 1.0000 0.88235294 0.3478261 1.0000000 0.6521739
#> 6  0.14705882 0.0000 0.85294118 1.0000 0.85294118 0.3555556 1.0000000 0.6444444
#> 7  0.17647059 0.0000 0.82352941 1.0000 0.82352941 0.3636364 1.0000000 0.6363636
#> 8  0.20588235 0.0000 0.79411765 1.0000 0.79411765 0.3720930 1.0000000 0.6279070
#> 9  0.23529412 0.0000 0.76470588 1.0000 0.76470588 0.3809524 1.0000000 0.6190476
#> 10 0.26470588 0.0000 0.73529412 1.0000 0.73529412 0.3902439 1.0000000 0.6097561
#> 11 0.29411765 0.0000 0.70588235 1.0000 0.70588235 0.4000000 1.0000000 0.6000000
#> 12 0.32352941 0.0000 0.67647059 1.0000 0.67647059 0.4102564 1.0000000 0.5897436
#> 13 0.35294118 0.0000 0.64705882 1.0000 0.64705882 0.4210526 1.0000000 0.5789474
#> 14 0.41176471 0.0000 0.58823529 1.0000 0.58823529 0.4444444 1.0000000 0.5555556
#> 15 0.44117647 0.0000 0.55882353 1.0000 0.55882353 0.4571429 1.0000000 0.5428571
#> 16 0.47058824 0.0000 0.52941176 1.0000 0.52941176 0.4705882 1.0000000 0.5294118
#> 17 0.50000000 0.0000 0.50000000 1.0000 0.50000000 0.4848485 1.0000000 0.5151515
#> 18 0.58823529 0.0000 0.41176471 1.0000 0.41176471 0.5333333 1.0000000 0.4666667
#> 19 0.61764706 0.0000 0.38235294 1.0000 0.38235294 0.5517241 1.0000000 0.4482759
#> 20 0.64705882 0.0000 0.35294118 1.0000 0.35294118 0.5714286 1.0000000 0.4285714
#> 21 0.67647059 0.0000 0.32352941 1.0000 0.32352941 0.5925926 1.0000000 0.4074074
#> 22 0.70588235 0.0000 0.29411765 1.0000 0.29411765 0.6153846 1.0000000 0.3846154
#> 23 0.76470588 0.0000 0.23529412 1.0000 0.23529412 0.6666667 1.0000000 0.3333333
#> 24 0.79411765 0.0000 0.20588235 1.0000 0.20588235 0.6956522 1.0000000 0.3043478
#> 25 0.82352941 0.0000 0.17647059 1.0000 0.17647059 0.7272727 1.0000000 0.2727273
#> 26 0.85294118 0.0000 0.14705882 1.0000 0.14705882 0.7619048 1.0000000 0.2380952
#> 27 0.85294118 0.0625 0.14705882 0.9375 0.14705882 0.7500000 0.9666667 0.2500000
#> 28 0.88235294 0.0625 0.11764706 0.9375 0.11764706 0.7894737 0.9677419 0.2105263
#> 29 0.88235294 0.1250 0.11764706 0.8750 0.11764706 0.7777778 0.9375000 0.2222222
#> 30 0.91176471 0.1250 0.08823529 0.8750 0.08823529 0.8235294 0.9393939 0.1764706
#> 31 0.91176471 0.1875 0.08823529 0.8125 0.08823529 0.8125000 0.9117647 0.1875000
#> 32 0.91176471 0.2500 0.08823529 0.7500 0.08823529 0.8000000 0.8857143 0.2000000
#> 33 0.94117647 0.2500 0.05882353 0.7500 0.05882353 0.8571429 0.8888889 0.1428571
#> 34 0.94117647 0.3750 0.05882353 0.6250 0.05882353 0.8333333 0.8421053 0.1666667
#> 35 0.94117647 0.4375 0.05882353 0.5625 0.05882353 0.8181818 0.8205128 0.1818182
#> 36 0.94117647 0.5000 0.05882353 0.5000 0.05882353 0.8000000 0.8000000 0.2000000
#> 37 0.94117647 0.5625 0.05882353 0.4375 0.05882353 0.7777778 0.7804878 0.2222222
#> 38 0.94117647 0.6250 0.05882353 0.3750 0.05882353 0.7500000 0.7619048 0.2500000
#> 39 0.94117647 0.6875 0.05882353 0.3125 0.05882353 0.7142857 0.7441860 0.2857143
#> 40 0.97058824 0.6875 0.02941176 0.3125 0.02941176 0.8333333 0.7500000 0.1666667
#> 41 0.97058824 0.7500 0.02941176 0.2500 0.02941176 0.8000000 0.7333333 0.2000000
#> 42 0.97058824 0.8125 0.02941176 0.1875 0.02941176 0.7500000 0.7173913 0.2500000
#> 43 1.00000000 0.8125 0.00000000 0.1875 0.00000000 1.0000000 0.7234043 0.0000000
#> 44 1.00000000 0.8750 0.00000000 0.1250 0.00000000 1.0000000 0.7083333 0.0000000
#> 45 1.00000000 0.9375 0.00000000 0.0625 0.00000000 1.0000000 0.6938776 0.0000000
#> 46 1.00000000 1.0000 0.00000000 0.0000 0.00000000        NA 0.6800000        NA
#>           FOR    youdenJ balancedAccuracy   f1Score        mcc
#> 1          NA 0.00000000        0.5000000 0.4848485         NA
#> 2  0.00000000 0.02941176        0.5147059 0.4923077 0.09799919
#> 3  0.00000000 0.05882353        0.5294118 0.5000000 0.14002801
#> 4  0.00000000 0.08823529        0.5441176 0.5079365 0.17331344
#> 5  0.00000000 0.11764706        0.5588235 0.5161290 0.20228869
#> 6  0.00000000 0.14705882        0.5735294 0.5245902 0.22866478
#> 7  0.00000000 0.17647059        0.5882353 0.5333333 0.25332020
#> 8  0.00000000 0.20588235        0.6029412 0.5423729 0.27678040
#> 9  0.00000000 0.23529412        0.6176471 0.5517241 0.29939248
#> 10 0.00000000 0.26470588        0.6323529 0.5614035 0.32140295
#> 11 0.00000000 0.29411765        0.6470588 0.5714286 0.34299717
#> 12 0.00000000 0.32352941        0.6617647 0.5818182 0.36432131
#> 13 0.00000000 0.35294118        0.6764706 0.5925926 0.38549554
#> 14 0.00000000 0.41176471        0.7058824 0.6153846 0.42779263
#> 15 0.00000000 0.44117647        0.7205882 0.6274510 0.44908871
#> 16 0.00000000 0.47058824        0.7352941 0.6400000 0.47058824
#> 17 0.00000000 0.50000000        0.7500000 0.6530612 0.49236596
#> 18 0.00000000 0.58823529        0.7941176 0.6956522 0.56011203
#> 19 0.00000000 0.61764706        0.8088235 0.7111111 0.58375576
#> 20 0.00000000 0.64705882        0.8235294 0.7272727 0.60806899
#> 21 0.00000000 0.67647059        0.8382353 0.7441860 0.63314411
#> 22 0.00000000 0.70588235        0.8529412 0.7619048 0.65908204
#> 23 0.00000000 0.76470588        0.8823529 0.8000000 0.71400555
#> 24 0.00000000 0.79411765        0.8970588 0.8205128 0.74325613
#> 25 0.00000000 0.82352941        0.9117647 0.8421053 0.77390599
#> 26 0.00000000 0.85294118        0.9264706 0.8648649 0.80613891
#> 27 0.03333333 0.79044118        0.8952206 0.8333333 0.75265055
#> 28 0.03225806 0.81985294        0.9099265 0.8571429 0.78791208
#> 29 0.06250000 0.75735294        0.8786765 0.8235294 0.73601476
#> 30 0.06060606 0.78676471        0.8933824 0.8484848 0.77475233
#> 31 0.08823529 0.72426471        0.8621324 0.8125000 0.72426471
#> 32 0.11428571 0.66176471        0.8308824 0.7741935 0.67363307
#> 33 0.11111111 0.69117647        0.8455882 0.8000000 0.71808049
#> 34 0.15789474 0.56617647        0.7830882 0.7142857 0.61839909
#> 35 0.17948718 0.50367647        0.7518382 0.6666667 0.56718204
#> 36 0.20000000 0.44117647        0.7205882 0.6153846 0.51449576
#> 37 0.21951220 0.37867647        0.6893382 0.5600000 0.45978478
#> 38 0.23809524 0.31617647        0.6580882 0.5000000 0.40230864
#> 39 0.25581395 0.25367647        0.6268382 0.4347826 0.34103299
#> 40 0.25000000 0.28308824        0.6415441 0.4545455 0.40636782
#> 41 0.26666667 0.22058824        0.6102941 0.3809524 0.34299717
#> 42 0.28260870 0.15808824        0.5790441 0.3000000 0.27182543
#> 43 0.27659574 0.18750000        0.5937500 0.3157895 0.36829105
#> 44 0.29166667 0.12500000        0.5625000 0.2222222 0.29755952
#> 45 0.30612245 0.06250000        0.5312500 0.1176471 0.20824828
#> 46 0.32000000 0.00000000        0.5000000        NA         NA
#>    diagnosticOddsRatio positiveLikelihoodRatio negativeLikelihoodRatio
#> 1                   NA                1.000000                      NA
#> 2                   NA                1.030303              0.00000000
#> 3                   NA                1.062500              0.00000000
#> 4                   NA                1.096774              0.00000000
#> 5                   NA                1.133333              0.00000000
#> 6                   NA                1.172414              0.00000000
#> 7                   NA                1.214286              0.00000000
#> 8                   NA                1.259259              0.00000000
#> 9                   NA                1.307692              0.00000000
#> 10                  NA                1.360000              0.00000000
#> 11                  NA                1.416667              0.00000000
#> 12                  NA                1.478261              0.00000000
#> 13                  NA                1.545455              0.00000000
#> 14                  NA                1.700000              0.00000000
#> 15                  NA                1.789474              0.00000000
#> 16                  NA                1.888889              0.00000000
#> 17                  NA                2.000000              0.00000000
#> 18                  NA                2.428571              0.00000000
#> 19                  NA                2.615385              0.00000000
#> 20                  NA                2.833333              0.00000000
#> 21                  NA                3.090909              0.00000000
#> 22                  NA                3.400000              0.00000000
#> 23                  NA                4.250000              0.00000000
#> 24                  NA                4.857143              0.00000000
#> 25                  NA                5.666667              0.00000000
#> 26                  NA                6.800000              0.00000000
#> 27           87.000000                6.375000              0.07327586
#> 28          112.500000                7.968750              0.07083333
#> 29           52.500000                7.437500              0.14166667
#> 30           72.333333                9.916667              0.13709677
#> 31           44.777778                9.208333              0.20564516
#> 32           31.000000                8.500000              0.27419355
#> 33           48.000000               12.750000              0.26562500
#> 34           26.666667               10.625000              0.39843750
#> 35           20.571429                9.562500              0.46484375
#> 36           16.000000                8.500000              0.53125000
#> 37           12.444444                7.437500              0.59765625
#> 38            9.600000                6.375000              0.66406250
#> 39            7.272727                5.312500              0.73046875
#> 40           15.000000               10.625000              0.70833333
#> 41           11.000000                8.500000              0.77272727
#> 42            7.615385                6.375000              0.83712121
#> 43                  NA                      NA              0.81250000
#> 44                  NA                      NA              0.87500000
#> 45                  NA                      NA              0.93750000
#> 46                  NA                      NA              1.00000000
#>    dPrimeSDT   betaSDT        cSDT      aSDT        bSDT
#> 1         NA        NA          NA        NA          NA
#> 2         NA 0.0000000          NA 0.7573529  0.02857143
#> 3         NA 0.0000000          NA 0.7647059  0.05555556
#> 4         NA 0.0000000          NA 0.7720588  0.08108108
#> 5         NA 0.0000000          NA 0.7794118  0.10526316
#> 6         NA 0.0000000          NA 0.7867647  0.12820513
#> 7         NA 0.0000000          NA 0.7941176  0.15000000
#> 8         NA 0.0000000          NA 0.8014706  0.17073171
#> 9         NA 0.0000000          NA 0.8088235  0.19047619
#> 10        NA 0.0000000          NA 0.8161765  0.20930233
#> 11        NA 0.0000000          NA 0.8235294  0.22727273
#> 12        NA 0.0000000          NA 0.8308824  0.24444444
#> 13        NA 0.0000000          NA 0.8382353  0.26086957
#> 14        NA 0.0000000          NA 0.8529412  0.29166667
#> 15        NA 0.0000000          NA 0.8602941  0.30612245
#> 16        NA 0.0000000          NA 0.8676471  0.32000000
#> 17        NA 0.0000000          NA 0.8750000  0.33333333
#> 18        NA 0.0000000          NA 0.8970588  0.37777778
#> 19        NA 0.0000000          NA 0.9044118  0.39534884
#> 20        NA 0.0000000          NA 0.9117647  0.41463415
#> 21        NA 0.0000000          NA 0.9191176  0.43589744
#> 22        NA 0.0000000          NA 0.9264706  0.45945946
#> 23        NA 0.0000000          NA 0.9411765  0.51515152
#> 24        NA 0.0000000          NA 0.9485294  0.54838710
#> 25        NA 0.0000000          NA 0.9558824  0.58620690
#> 26        NA 0.0000000          NA 0.9632353  0.62962963
#> 27  2.583252 0.5344995 -0.24249457 0.9384191  0.78703704
#> 28  2.720952 0.6234551 -0.17364456 0.9476103  0.85000000
#> 29  2.337181 1.0435544  0.01824103 0.9246324  1.02000000
#> 30  2.502052 1.2864643  0.10067643 0.9356618  1.10869565
#> 31  2.238849 1.6820865  0.23227784 0.9145221  1.29347826
#> 32  2.026192 1.9859180  0.33860624 0.8933824  1.47826087
#> 33  2.239216 2.7093703  0.44511836 0.9080882  1.61904762
#> 34  1.883366 3.2330273  0.62304355 0.8694853  2.02380952
#> 35  1.722037 3.3595638  0.70370789 0.8501838  2.22619048
#> 36  1.564726 3.4013910  0.78236324 0.8308824  2.42857143
#> 37  1.407416 3.3595638  0.86101858 0.8110557  2.51331497
#> 38  1.246087 3.2330273  0.94168292 0.7898284  2.58525346
#> 39  1.075950 3.0184193  1.02675144 0.7663603  2.62114537
#> 40  1.400734 5.2892304  1.18914319 0.7972426  3.22784810
#> 41  1.215020 4.7476784  1.28199986 0.7757353  3.40000000
#> 42  1.002363 4.0213169  1.38832826 0.7503064  3.44839858
#> 43        NA        NA          NA 0.7968750  6.33333333
#> 44        NA        NA          NA 0.7812500  9.00000000
#> 45        NA        NA          NA 0.7656250 17.00000000
#> 46        NA        NA          NA        NA          NA
#>    differenceBetweenPredictedAndObserved informationGain overallUtility
#> 1                                49.6000              NA          0.320
#> 2                                49.6000              NA          0.338
#> 3                                49.6000              NA          0.356
#> 4                                49.6000              NA          0.374
#> 5                                49.6000              NA          0.392
#> 6                                75.8000              NA          0.410
#> 7                                75.8000              NA          0.428
#> 8                                75.8000              NA          0.446
#> 9                                75.8000              NA          0.464
#> 10                               75.8000              NA          0.482
#> 11                              107.2000              NA          0.500
#> 12                              107.2000              NA          0.518
#> 13                              107.2000              NA          0.536
#> 14                              107.2000              NA          0.572
#> 15                              117.6000              NA          0.590
#> 16                              117.6000              NA          0.608
#> 17                              117.6000              NA          0.626
#> 18                              153.1667              NA          0.680
#> 19                              153.1667              NA          0.698
#> 20                              153.1667              NA          0.716
#> 21                              153.1667              NA          0.734
#> 22                              175.0000              NA          0.752
#> 23                              175.0000              NA          0.788
#> 24                              175.0000              NA          0.806
#> 25                              175.0000              NA          0.824
#> 26                              175.0000              NA          0.842
#> 27                              207.8000      0.45336483          0.822
#> 28                              207.8000      0.49476883          0.840
#> 29                              207.8000      0.41340219          0.820
#> 30                              207.8000      0.45810281          0.838
#> 31                              207.8000      0.38881826          0.818
#> 32                              247.6000      0.32890663          0.798
#> 33                              247.6000      0.37636708          0.816
#> 34                              247.6000      0.27014667          0.776
#> 35                              265.6000      0.22430896          0.756
#> 36                              265.6000      0.18245336          0.736
#> 37                              265.6000      0.14421850          0.716
#> 38                              265.6000      0.10941594          0.696
#> 39                              265.6000      0.07803185          0.676
#> 40                              265.6000      0.11245402          0.694
#> 41                              309.4000      0.07921198          0.674
#> 42                              309.4000      0.04921665          0.654
#> 43                              309.4000              NA          0.672
#> 44                              309.4000              NA          0.652
#> 45                              309.4000              NA          0.632
#> 46                                    NA              NA          0.612