Overall Accuracy.

Find overall accuracy.

Usage

accuracyOverall(predicted, actual, dropUndefined = FALSE)

Arguments

predicted

vector of continuous predicted values.

actual

vector of actual values.

dropUndefined

TRUE or FALSE, indicating whether to drop any undefined values calculated with the accuracy indices.

Value

ME = mean error

MAE = mean absolute error

MSE = mean squared error

RMSE = root mean squared error

MPE = mean percentage error

MAPE = mean absolute percentage error

sMAPE = symmetric mean absolute percentage error

MASE = mean absolute scaled error

RMSLE = root mean squared log error

rsquared = R-squared

rsquaredAdj = adjusted R-squared

rsquaredPredictive = predictive R-squared

Details

Compute overall accuracy estimates of predicted values in relation to actual values. Estimates of overall accuracy span all cutoffs. Some accuracy estimates can be undefined under various circumstances. Optionally, you can drop undefined values in the calculation of accuracy indices. Note that dropping undefined values changes the meaning of these indices. Use this option at your own risk!

See also

Examples

# Prepare Data
data("USArrests")

# Calculate Accuracy
accuracyOverall(predicted = USArrests$Assault, actual = USArrests$Murder)
#>        ME     MAE      MSE    RMSE       MPE     MAPE    sMAPE    MASE    RMSLE
#> 1 162.972 162.972 32814.24 181.147 -2382.823 2382.823 91.16591 44.6861 2.990233
#>    rsquared rsquaredAdj rsquaredPredictive
#> 1 0.6430008   0.6355633          0.6163406