Package Dependencies.

Determine package dependencies.

Usage

getDependencies(packs)

Arguments

packs

Character vector of names of target packages.

Value

Vector of packages that depend on the target package(s).

Details

Determine which packages depend on a target package (or packages).

See also

Examples

old <- options("repos")
options(repos = "https://cran.r-project.org")
getDependencies("tidyverse")
#>   [1] "tidyverse"     "broom"         "conflicted"    "cli"          
#>   [5] "dbplyr"        "dplyr"         "dtplyr"        "forcats"      
#>   [9] "ggplot2"       "googledrive"   "googlesheets4" "haven"        
#>  [13] "hms"           "httr"          "jsonlite"      "lubridate"    
#>  [17] "magrittr"      "modelr"        "pillar"        "purrr"        
#>  [21] "ragg"          "readr"         "readxl"        "reprex"       
#>  [25] "rlang"         "rstudioapi"    "rvest"         "stringr"      
#>  [29] "tibble"        "tidyr"         "xml2"          "backports"    
#>  [33] "generics"      "glue"          "lifecycle"     "utils"        
#>  [37] "memoise"       "blob"          "DBI"           "methods"      
#>  [41] "R6"            "tidyselect"    "vctrs"         "withr"        
#>  [45] "data.table"    "grDevices"     "grid"          "gtable"       
#>  [49] "isoband"       "MASS"          "mgcv"          "scales"       
#>  [53] "stats"         "gargle"        "uuid"          "cellranger"   
#>  [57] "curl"          "ids"           "rematch2"      "pkgconfig"    
#>  [61] "mime"          "openssl"       "timechange"    "fansi"        
#>  [65] "utf8"          "systemfonts"   "textshaping"   "clipr"        
#>  [69] "crayon"        "vroom"         "callr"         "fs"           
#>  [73] "knitr"         "rmarkdown"     "selectr"       "stringi"      
#>  [77] "graphics"      "processx"      "rematch"       "rappdirs"     
#>  [81] "evaluate"      "highr"         "tools"         "xfun"         
#>  [85] "yaml"          "cachem"        "nlme"          "Matrix"       
#>  [89] "splines"       "askpass"       "bslib"         "fontawesome"  
#>  [93] "htmltools"     "jquerylib"     "tinytex"       "farver"       
#>  [97] "labeling"      "munsell"       "RColorBrewer"  "viridisLite"  
#> [101] "bit64"         "tzdb"          "lattice"       "sys"          
#> [105] "bit"           "base64enc"     "fastmap"       "sass"         
#> [109] "digest"        "colorspace"    "ps"           
options(old)