Proportion of Maximum (POM).

Calculate the proportion of maximum (POM) score given a minimum and maximum score.

Usage

pom(data, min = NULL, max = NULL)

Arguments

data

The vector of data.

min

The minimum possible or observed value.

max

The maximum possible or observed value.

Value

Proportion of maximum possible or observed values.

Details

The minimum and maximum score for calculating the proportion of maximum could be the possible or observed minimum and maximum, respectively. Using the possible minimum and maximum would yield the proportion of maximum possible score. Using the observed minimum and maximum would yield the proportion of minimum and maximum observed score. If the minimum and maximum possible scores are not specified, the observed minimum and maximum are used.

See also

Examples

# Prepare Data
v1 <- sample(1:9, size = 1000, replace = TRUE)

# Calculate Proportion of Maximum Possible (by specifying the minimum and maximum possible)
pom(v1, min = 0, max = 10)
#>    [1] 0.6 0.7 0.1 0.7 0.1 0.6 0.7 0.7 0.8 0.5 0.4 0.2 0.7 0.2 0.8 0.5 0.7 0.8
#>   [19] 0.8 0.7 0.4 0.9 0.2 0.2 0.4 0.4 0.9 0.9 0.1 0.9 0.7 0.5 0.9 0.2 0.8 0.6
#>   [37] 0.9 0.2 0.8 0.6 0.4 0.7 0.9 0.6 0.8 0.1 0.3 0.2 0.5 0.9 0.9 0.5 0.9 0.4
#>   [55] 0.6 0.3 0.3 0.3 0.8 0.1 0.6 0.6 0.4 0.6 0.6 0.5 0.1 0.7 0.8 0.9 0.7 0.1
#>   [73] 0.4 0.7 0.5 0.9 0.5 0.6 0.6 0.4 0.8 0.6 0.3 0.4 0.8 0.7 0.5 0.6 0.8 0.9
#>   [91] 0.7 0.8 0.2 0.8 0.2 0.5 0.4 0.2 0.6 0.5 0.8 0.4 0.7 0.8 0.3 0.7 0.7 0.2
#>  [109] 0.3 0.8 0.5 0.9 0.1 0.9 0.3 0.5 0.3 0.4 0.1 0.3 0.9 0.9 0.8 0.7 0.8 0.6
#>  [127] 0.1 0.7 0.9 0.9 0.1 0.7 0.1 0.2 0.1 0.7 0.7 0.1 0.7 0.6 0.6 0.1 0.5 0.9
#>  [145] 0.6 0.4 0.4 0.7 0.9 0.3 0.4 0.1 0.7 0.8 0.8 0.6 0.8 0.8 0.8 0.2 0.1 0.7
#>  [163] 0.4 0.9 0.7 0.6 0.4 0.8 0.2 0.3 0.4 0.5 0.2 0.5 0.7 0.9 0.5 0.6 0.6 0.3
#>  [181] 0.3 0.3 0.9 0.4 0.1 0.3 0.9 0.8 0.5 0.5 0.2 0.1 0.5 0.9 0.6 0.3 0.3 0.1
#>  [199] 0.3 0.4 0.4 0.7 0.5 0.5 0.6 0.3 0.4 0.4 0.8 0.4 0.5 0.9 0.4 0.1 0.3 0.7
#>  [217] 0.9 0.6 0.3 0.1 0.4 0.2 0.8 0.1 0.4 0.5 0.3 0.6 0.5 0.1 0.4 0.5 0.3 0.2
#>  [235] 0.8 0.1 0.4 0.1 0.9 0.3 0.8 0.7 0.6 0.8 0.5 0.5 0.4 0.2 0.7 0.3 0.1 0.5
#>  [253] 0.8 0.4 0.1 0.5 0.8 0.5 0.6 0.8 0.7 0.7 0.6 0.5 0.4 0.9 0.8 0.9 0.6 0.7
#>  [271] 0.6 0.1 0.3 0.1 0.4 0.4 0.9 0.4 0.6 0.4 0.3 0.7 0.9 0.9 0.7 0.7 0.4 0.9
#>  [289] 0.2 0.8 0.9 0.7 0.5 0.1 0.9 0.5 0.7 0.8 0.7 0.6 0.9 0.2 0.7 0.7 0.5 0.3
#>  [307] 0.8 0.5 0.9 0.5 0.1 0.8 0.6 0.6 0.4 0.1 0.3 0.7 0.2 0.4 0.1 0.2 0.9 0.4
#>  [325] 0.1 0.2 0.6 0.2 0.6 0.7 0.6 0.9 0.9 0.7 0.8 0.1 0.1 0.4 0.5 0.5 0.4 0.7
#>  [343] 0.8 0.1 0.3 0.2 0.4 0.7 0.8 0.7 0.5 0.1 0.9 0.1 0.4 0.4 0.4 0.2 0.8 0.5
#>  [361] 0.7 0.7 0.2 0.9 0.3 0.3 0.8 0.2 0.3 0.4 0.3 0.6 0.3 0.9 0.8 0.2 0.4 0.7
#>  [379] 0.6 0.4 0.1 0.5 0.5 0.6 0.8 0.3 0.2 0.6 0.4 0.4 0.2 0.3 0.9 0.1 0.1 0.4
#>  [397] 0.6 0.1 0.5 0.2 0.5 0.2 0.6 0.9 0.4 0.2 0.6 0.2 0.1 0.9 0.8 0.7 0.3 0.3
#>  [415] 0.9 0.7 0.1 0.4 0.7 0.3 0.6 0.9 0.7 0.9 0.6 0.2 0.9 0.9 0.2 0.1 0.8 0.4
#>  [433] 0.9 0.2 0.4 0.6 0.8 0.3 0.1 0.2 0.4 0.5 0.5 0.6 0.1 0.7 0.9 0.6 0.7 0.5
#>  [451] 0.8 0.9 0.2 0.6 0.4 0.4 0.4 0.4 0.2 0.3 0.3 0.8 0.6 0.8 0.2 0.8 0.7 0.8
#>  [469] 0.4 0.7 0.8 0.3 0.4 0.7 0.5 0.8 0.5 0.7 0.9 0.2 0.6 0.1 0.5 0.5 0.7 0.6
#>  [487] 0.5 0.3 0.9 0.3 0.3 0.5 0.7 0.3 0.5 0.3 0.2 0.4 0.6 0.6 0.7 0.4 0.9 0.9
#>  [505] 0.4 0.3 0.3 0.3 0.3 0.9 0.3 0.6 0.9 0.8 0.7 0.7 0.5 0.3 0.2 0.9 0.7 0.9
#>  [523] 0.9 0.7 0.1 0.7 0.2 0.5 0.3 0.4 0.6 0.6 0.7 0.4 0.7 0.3 0.4 0.4 0.4 0.7
#>  [541] 0.5 0.1 0.2 0.2 0.7 0.9 0.6 0.6 0.4 0.4 0.4 0.6 0.9 0.9 0.8 0.2 0.6 0.1
#>  [559] 0.3 0.5 0.6 0.6 0.1 0.7 0.5 0.6 0.4 0.1 0.1 0.5 0.6 0.7 0.5 0.2 0.8 0.3
#>  [577] 0.5 0.7 0.3 0.7 0.2 0.7 0.8 0.8 0.6 0.3 0.1 0.5 0.3 0.3 0.6 0.1 0.8 0.1
#>  [595] 0.1 0.3 0.4 0.9 0.4 0.3 0.4 0.5 0.2 0.1 0.8 0.8 0.3 0.6 0.2 0.8 0.5 0.1
#>  [613] 0.1 0.3 0.7 0.2 0.2 0.4 0.5 0.3 0.1 0.6 0.3 0.5 0.5 0.1 0.6 0.2 0.3 0.4
#>  [631] 0.4 0.4 0.9 0.1 0.6 0.8 0.7 0.4 0.8 0.4 0.3 0.4 0.2 0.1 0.6 0.2 0.9 0.4
#>  [649] 0.6 0.8 0.8 0.6 0.1 0.9 0.7 0.2 0.8 0.9 0.9 0.2 0.7 0.2 0.7 0.4 0.5 0.8
#>  [667] 0.7 0.2 0.8 0.9 0.1 0.4 0.2 0.1 0.6 0.6 0.3 0.2 0.5 0.2 0.5 0.5 0.4 0.1
#>  [685] 0.6 0.3 0.5 0.8 0.2 0.7 0.6 0.6 0.5 0.1 0.5 0.5 0.1 0.6 0.5 0.2 0.7 0.8
#>  [703] 0.7 0.7 0.5 0.5 0.5 0.3 0.4 0.6 0.5 0.4 0.6 0.1 0.3 0.5 0.8 0.1 0.7 0.4
#>  [721] 0.8 0.1 0.7 0.3 0.1 0.4 0.5 0.6 0.4 0.2 0.7 0.3 0.8 0.1 0.8 0.4 0.7 0.5
#>  [739] 0.1 0.2 0.3 0.6 0.3 0.7 0.4 0.2 0.2 0.6 0.4 0.6 0.5 0.3 0.3 0.3 0.3 0.6
#>  [757] 0.1 0.7 0.1 0.1 0.6 0.8 0.3 0.6 0.9 0.7 0.8 0.9 0.3 0.7 0.2 0.9 0.9 0.6
#>  [775] 0.7 0.3 0.1 0.6 0.8 0.1 0.4 0.3 0.6 0.9 0.6 0.8 0.5 0.4 0.2 0.8 0.9 0.2
#>  [793] 0.9 0.2 0.7 0.1 0.6 0.9 0.7 0.2 0.2 0.7 0.6 0.9 0.7 0.3 0.9 0.8 0.5 0.8
#>  [811] 0.3 0.2 0.1 0.8 0.2 0.1 0.2 0.1 0.9 0.1 0.2 0.3 0.3 0.8 0.7 0.3 0.5 0.5
#>  [829] 0.6 0.2 0.3 0.2 0.5 0.3 0.9 0.1 0.7 0.8 0.6 0.9 0.5 0.4 0.4 0.3 0.8 0.1
#>  [847] 0.8 0.7 0.3 0.4 0.6 0.5 0.4 0.7 0.8 0.7 0.3 0.3 0.5 0.5 0.1 0.1 0.8 0.5
#>  [865] 0.5 0.7 0.2 0.2 0.1 0.3 0.5 0.9 0.6 0.8 0.7 0.9 0.9 0.9 0.2 0.4 0.7 0.9
#>  [883] 0.6 0.5 0.6 0.2 0.4 0.5 0.7 0.7 0.3 0.7 0.7 0.1 0.7 0.2 0.6 0.1 0.9 0.7
#>  [901] 0.9 0.2 0.4 0.3 0.1 0.7 0.5 0.4 0.9 0.5 0.1 0.3 0.7 0.5 0.3 0.8 0.1 0.5
#>  [919] 0.3 0.3 0.8 0.4 0.8 0.4 0.8 0.7 0.7 0.7 0.2 0.3 0.2 0.3 0.7 0.2 0.9 0.2
#>  [937] 0.7 0.5 0.4 0.2 0.6 0.1 0.6 0.5 0.9 0.7 0.6 0.9 0.7 0.8 0.4 0.7 0.3 0.3
#>  [955] 0.4 0.1 0.4 0.5 0.2 0.6 0.3 0.5 0.7 0.2 0.9 0.5 0.9 0.1 0.3 0.1 0.3 0.8
#>  [973] 0.8 0.4 0.5 0.1 0.1 0.2 0.6 0.9 0.8 0.9 0.1 0.9 0.2 0.4 0.1 0.3 0.3 0.4
#>  [991] 0.5 0.6 0.9 0.3 0.1 0.2 0.2 0.1 0.8 0.5

# Calculate Proportion of Maximum Observed
pom(v1)
#>    [1] 0.625 0.750 0.000 0.750 0.000 0.625 0.750 0.750 0.875 0.500 0.375 0.125
#>   [13] 0.750 0.125 0.875 0.500 0.750 0.875 0.875 0.750 0.375 1.000 0.125 0.125
#>   [25] 0.375 0.375 1.000 1.000 0.000 1.000 0.750 0.500 1.000 0.125 0.875 0.625
#>   [37] 1.000 0.125 0.875 0.625 0.375 0.750 1.000 0.625 0.875 0.000 0.250 0.125
#>   [49] 0.500 1.000 1.000 0.500 1.000 0.375 0.625 0.250 0.250 0.250 0.875 0.000
#>   [61] 0.625 0.625 0.375 0.625 0.625 0.500 0.000 0.750 0.875 1.000 0.750 0.000
#>   [73] 0.375 0.750 0.500 1.000 0.500 0.625 0.625 0.375 0.875 0.625 0.250 0.375
#>   [85] 0.875 0.750 0.500 0.625 0.875 1.000 0.750 0.875 0.125 0.875 0.125 0.500
#>   [97] 0.375 0.125 0.625 0.500 0.875 0.375 0.750 0.875 0.250 0.750 0.750 0.125
#>  [109] 0.250 0.875 0.500 1.000 0.000 1.000 0.250 0.500 0.250 0.375 0.000 0.250
#>  [121] 1.000 1.000 0.875 0.750 0.875 0.625 0.000 0.750 1.000 1.000 0.000 0.750
#>  [133] 0.000 0.125 0.000 0.750 0.750 0.000 0.750 0.625 0.625 0.000 0.500 1.000
#>  [145] 0.625 0.375 0.375 0.750 1.000 0.250 0.375 0.000 0.750 0.875 0.875 0.625
#>  [157] 0.875 0.875 0.875 0.125 0.000 0.750 0.375 1.000 0.750 0.625 0.375 0.875
#>  [169] 0.125 0.250 0.375 0.500 0.125 0.500 0.750 1.000 0.500 0.625 0.625 0.250
#>  [181] 0.250 0.250 1.000 0.375 0.000 0.250 1.000 0.875 0.500 0.500 0.125 0.000
#>  [193] 0.500 1.000 0.625 0.250 0.250 0.000 0.250 0.375 0.375 0.750 0.500 0.500
#>  [205] 0.625 0.250 0.375 0.375 0.875 0.375 0.500 1.000 0.375 0.000 0.250 0.750
#>  [217] 1.000 0.625 0.250 0.000 0.375 0.125 0.875 0.000 0.375 0.500 0.250 0.625
#>  [229] 0.500 0.000 0.375 0.500 0.250 0.125 0.875 0.000 0.375 0.000 1.000 0.250
#>  [241] 0.875 0.750 0.625 0.875 0.500 0.500 0.375 0.125 0.750 0.250 0.000 0.500
#>  [253] 0.875 0.375 0.000 0.500 0.875 0.500 0.625 0.875 0.750 0.750 0.625 0.500
#>  [265] 0.375 1.000 0.875 1.000 0.625 0.750 0.625 0.000 0.250 0.000 0.375 0.375
#>  [277] 1.000 0.375 0.625 0.375 0.250 0.750 1.000 1.000 0.750 0.750 0.375 1.000
#>  [289] 0.125 0.875 1.000 0.750 0.500 0.000 1.000 0.500 0.750 0.875 0.750 0.625
#>  [301] 1.000 0.125 0.750 0.750 0.500 0.250 0.875 0.500 1.000 0.500 0.000 0.875
#>  [313] 0.625 0.625 0.375 0.000 0.250 0.750 0.125 0.375 0.000 0.125 1.000 0.375
#>  [325] 0.000 0.125 0.625 0.125 0.625 0.750 0.625 1.000 1.000 0.750 0.875 0.000
#>  [337] 0.000 0.375 0.500 0.500 0.375 0.750 0.875 0.000 0.250 0.125 0.375 0.750
#>  [349] 0.875 0.750 0.500 0.000 1.000 0.000 0.375 0.375 0.375 0.125 0.875 0.500
#>  [361] 0.750 0.750 0.125 1.000 0.250 0.250 0.875 0.125 0.250 0.375 0.250 0.625
#>  [373] 0.250 1.000 0.875 0.125 0.375 0.750 0.625 0.375 0.000 0.500 0.500 0.625
#>  [385] 0.875 0.250 0.125 0.625 0.375 0.375 0.125 0.250 1.000 0.000 0.000 0.375
#>  [397] 0.625 0.000 0.500 0.125 0.500 0.125 0.625 1.000 0.375 0.125 0.625 0.125
#>  [409] 0.000 1.000 0.875 0.750 0.250 0.250 1.000 0.750 0.000 0.375 0.750 0.250
#>  [421] 0.625 1.000 0.750 1.000 0.625 0.125 1.000 1.000 0.125 0.000 0.875 0.375
#>  [433] 1.000 0.125 0.375 0.625 0.875 0.250 0.000 0.125 0.375 0.500 0.500 0.625
#>  [445] 0.000 0.750 1.000 0.625 0.750 0.500 0.875 1.000 0.125 0.625 0.375 0.375
#>  [457] 0.375 0.375 0.125 0.250 0.250 0.875 0.625 0.875 0.125 0.875 0.750 0.875
#>  [469] 0.375 0.750 0.875 0.250 0.375 0.750 0.500 0.875 0.500 0.750 1.000 0.125
#>  [481] 0.625 0.000 0.500 0.500 0.750 0.625 0.500 0.250 1.000 0.250 0.250 0.500
#>  [493] 0.750 0.250 0.500 0.250 0.125 0.375 0.625 0.625 0.750 0.375 1.000 1.000
#>  [505] 0.375 0.250 0.250 0.250 0.250 1.000 0.250 0.625 1.000 0.875 0.750 0.750
#>  [517] 0.500 0.250 0.125 1.000 0.750 1.000 1.000 0.750 0.000 0.750 0.125 0.500
#>  [529] 0.250 0.375 0.625 0.625 0.750 0.375 0.750 0.250 0.375 0.375 0.375 0.750
#>  [541] 0.500 0.000 0.125 0.125 0.750 1.000 0.625 0.625 0.375 0.375 0.375 0.625
#>  [553] 1.000 1.000 0.875 0.125 0.625 0.000 0.250 0.500 0.625 0.625 0.000 0.750
#>  [565] 0.500 0.625 0.375 0.000 0.000 0.500 0.625 0.750 0.500 0.125 0.875 0.250
#>  [577] 0.500 0.750 0.250 0.750 0.125 0.750 0.875 0.875 0.625 0.250 0.000 0.500
#>  [589] 0.250 0.250 0.625 0.000 0.875 0.000 0.000 0.250 0.375 1.000 0.375 0.250
#>  [601] 0.375 0.500 0.125 0.000 0.875 0.875 0.250 0.625 0.125 0.875 0.500 0.000
#>  [613] 0.000 0.250 0.750 0.125 0.125 0.375 0.500 0.250 0.000 0.625 0.250 0.500
#>  [625] 0.500 0.000 0.625 0.125 0.250 0.375 0.375 0.375 1.000 0.000 0.625 0.875
#>  [637] 0.750 0.375 0.875 0.375 0.250 0.375 0.125 0.000 0.625 0.125 1.000 0.375
#>  [649] 0.625 0.875 0.875 0.625 0.000 1.000 0.750 0.125 0.875 1.000 1.000 0.125
#>  [661] 0.750 0.125 0.750 0.375 0.500 0.875 0.750 0.125 0.875 1.000 0.000 0.375
#>  [673] 0.125 0.000 0.625 0.625 0.250 0.125 0.500 0.125 0.500 0.500 0.375 0.000
#>  [685] 0.625 0.250 0.500 0.875 0.125 0.750 0.625 0.625 0.500 0.000 0.500 0.500
#>  [697] 0.000 0.625 0.500 0.125 0.750 0.875 0.750 0.750 0.500 0.500 0.500 0.250
#>  [709] 0.375 0.625 0.500 0.375 0.625 0.000 0.250 0.500 0.875 0.000 0.750 0.375
#>  [721] 0.875 0.000 0.750 0.250 0.000 0.375 0.500 0.625 0.375 0.125 0.750 0.250
#>  [733] 0.875 0.000 0.875 0.375 0.750 0.500 0.000 0.125 0.250 0.625 0.250 0.750
#>  [745] 0.375 0.125 0.125 0.625 0.375 0.625 0.500 0.250 0.250 0.250 0.250 0.625
#>  [757] 0.000 0.750 0.000 0.000 0.625 0.875 0.250 0.625 1.000 0.750 0.875 1.000
#>  [769] 0.250 0.750 0.125 1.000 1.000 0.625 0.750 0.250 0.000 0.625 0.875 0.000
#>  [781] 0.375 0.250 0.625 1.000 0.625 0.875 0.500 0.375 0.125 0.875 1.000 0.125
#>  [793] 1.000 0.125 0.750 0.000 0.625 1.000 0.750 0.125 0.125 0.750 0.625 1.000
#>  [805] 0.750 0.250 1.000 0.875 0.500 0.875 0.250 0.125 0.000 0.875 0.125 0.000
#>  [817] 0.125 0.000 1.000 0.000 0.125 0.250 0.250 0.875 0.750 0.250 0.500 0.500
#>  [829] 0.625 0.125 0.250 0.125 0.500 0.250 1.000 0.000 0.750 0.875 0.625 1.000
#>  [841] 0.500 0.375 0.375 0.250 0.875 0.000 0.875 0.750 0.250 0.375 0.625 0.500
#>  [853] 0.375 0.750 0.875 0.750 0.250 0.250 0.500 0.500 0.000 0.000 0.875 0.500
#>  [865] 0.500 0.750 0.125 0.125 0.000 0.250 0.500 1.000 0.625 0.875 0.750 1.000
#>  [877] 1.000 1.000 0.125 0.375 0.750 1.000 0.625 0.500 0.625 0.125 0.375 0.500
#>  [889] 0.750 0.750 0.250 0.750 0.750 0.000 0.750 0.125 0.625 0.000 1.000 0.750
#>  [901] 1.000 0.125 0.375 0.250 0.000 0.750 0.500 0.375 1.000 0.500 0.000 0.250
#>  [913] 0.750 0.500 0.250 0.875 0.000 0.500 0.250 0.250 0.875 0.375 0.875 0.375
#>  [925] 0.875 0.750 0.750 0.750 0.125 0.250 0.125 0.250 0.750 0.125 1.000 0.125
#>  [937] 0.750 0.500 0.375 0.125 0.625 0.000 0.625 0.500 1.000 0.750 0.625 1.000
#>  [949] 0.750 0.875 0.375 0.750 0.250 0.250 0.375 0.000 0.375 0.500 0.125 0.625
#>  [961] 0.250 0.500 0.750 0.125 1.000 0.500 1.000 0.000 0.250 0.000 0.250 0.875
#>  [973] 0.875 0.375 0.500 0.000 0.000 0.125 0.625 1.000 0.875 1.000 0.000 1.000
#>  [985] 0.125 0.375 0.000 0.250 0.250 0.375 0.500 0.625 1.000 0.250 0.000 0.125
#>  [997] 0.125 0.000 0.875 0.500