includes/google_analytics.html

Read Encrypted Data.

Read data from encrypted file.

Usage

read.aes(filename, key)

Arguments

filename

Location of encrypted data.

key

Encryption key.

Value

Unencrypted data.

Details

Reads data from an encrypted file. To write an data to an encrypted file, see write.aes.

See also

Examples

# Location of Encryption Key on Local Computer (where only you should have access to it)
encryptionKeyLocation <- file.path(getwd(), "/encryptionKey.RData",
  fsep = "") #Can change to a different path, e.g.: "C:/Users/[USERNAME]/"

# Generate Encryption Key
key <- as.raw(sample(1:16, 16))

# Save Encryption Key
save(key, file = encryptionKeyLocation)

# Specify Credentials
credentials <- "Insert My Credentials Here"

# Save Encrypted Credentials
write.aes(df = credentials,
  filename = file.path(getwd(), "/encrypytedCredentials.txt", fsep = ""),
  key = key) #Change the file location to save this on the lab drive
rm(credentials)
rm(key)

# Read and Unencrypt the Credentials Using the Encryption Key
load(encryptionKeyLocation)
credentials <- read.aes(filename =
  file.path(getwd(), "/encrypytedCredentials.txt", fsep = ""), key = key)

includes/footer.html