Drop NA rows.

Drop rows with all missing (NA) values.

Usage

dropRowsWithAllNA(data, ignore = NULL)

Arguments

data

Dataframe to drop rows from.

ignore

Names of columns to ignore for determining whether each row had all missing values.

Value

A dataframe with rows removed that had all missing values in non-ignored columns.

Details

Drop rows that have no observed values, i.e., all values in the row are missing (NA), excluding the ignored columns.

See also

Examples

# Prepare Data
df <- expand.grid(ID = 1:100, time = c(1, 2, 3))
df <- df[order(df$ID),]
row.names(df) <- NULL
df$score1 <- rnorm(nrow(df))
df$score2 <- rnorm(nrow(df))
df$score3 <- rnorm(nrow(df))
df[sample(1:nrow(df), size = 100), c("score1","score2","score3")] <- NA

# Drop Rows with All NA in Non-Ignored Columns
dropRowsWithAllNA(df, ignore = c("ID","time"))
#>      ID time      score1      score2       score3
#> 1     1    1 -0.41045598  1.23407764 -0.551239629
#> 4     2    1  0.79497460  0.49528606 -0.924742724
#> 5     2    2  0.27329947 -0.53403248 -0.345142313
#> 6     2    3  0.90589274  1.20949738 -0.823168807
#> 8     3    2  0.27353905  0.48360543  0.700960895
#> 10    4    1  1.66813999  0.99245161  1.139158270
#> 11    4    2  0.45086048 -0.73042160  0.170783997
#> 12    4    3 -0.56360706  0.63374342 -0.007159302
#> 14    5    2 -0.64484901  0.53540384 -0.214636639
#> 15    5    3  1.67482941 -0.50840230 -0.633826615
#> 16    6    1  1.25631456  1.95536073  0.013072450
#> 18    6    3 -0.30634707  1.43845252 -0.444209491
#> 19    7    1 -1.57748241 -0.32589152  0.220425553
#> 20    7    2  0.08715505 -1.02519855  0.638901159
#> 21    7    3 -1.10674648 -0.04390200  0.580394014
#> 25    9    1 -0.88035929  0.43564150  0.330066964
#> 26    9    2  1.53642213  2.04509405 -0.657893957
#> 28   10    1 -1.54947961  0.72410853  1.200836500
#> 29   10    2 -1.68825272  0.20801872 -0.300812402
#> 33   11    3  0.15447805  1.68277464  2.343104441
#> 34   12    1 -0.32082590  1.96927191 -0.177075297
#> 35   12    2  1.56296947  0.50153864 -0.043803878
#> 37   13    1 -1.00629519  0.17757223  2.497815829
#> 38   13    2 -0.18346959 -0.72381599 -0.634933486
#> 40   14    1  0.51908829 -1.60657709  0.406992525
#> 41   14    2  0.44910444  2.09277438  1.103273907
#> 43   15    1  0.50620016  0.15861449  0.265610157
#> 44   15    2 -0.47012106  1.81618883  0.163216140
#> 45   15    3  1.05272066  0.42402218 -0.934675712
#> 46   16    1 -0.54825561  0.75353450  0.518930279
#> 47   16    2 -1.81571663  1.28912411 -0.158326401
#> 48   16    3 -1.02502449  1.02151744  0.535134099
#> 49   17    1 -1.50541956  0.34032502 -1.286095649
#> 50   17    2  1.14848016  0.10249559  1.039636333
#> 51   17    3 -0.23544628 -0.92318861 -0.220851435
#> 52   18    1  0.98051147  1.84860068  0.707972016
#> 53   18    2 -0.68132670  0.17954935 -0.072282608
#> 54   18    3 -0.53528570  0.80626799 -0.196291321
#> 57   19    3  1.32052128  1.48203488  0.764619070
#> 58   20    1 -0.55356393  0.03561794  1.235965907
#> 59   20    2  1.35071353 -0.22886523 -0.455143131
#> 60   20    3 -0.80623665 -1.14973765  1.463705935
#> 63   21    3  1.43469619  1.16813576  1.315313346
#> 65   22    2 -0.06673113 -1.04967722  1.219572694
#> 66   22    3 -0.90622583  0.66986085 -2.392030396
#> 67   23    1 -0.40659996  1.40121241  0.439299875
#> 68   23    2 -0.68556722  0.52342579 -0.311222883
#> 69   23    3  0.95667584 -0.31820181 -0.199988863
#> 70   24    1 -0.22691983  1.12657695  1.662453852
#> 71   24    2 -1.12186180 -0.36408292 -0.348245415
#> 72   24    3  0.41725700 -0.78366728  0.866831654
#> 74   25    2  0.29196048 -1.06471645  0.989521227
#> 75   25    3 -0.79535161  0.93839510  0.934002586
#> 76   26    1 -0.62924534 -0.66755520 -0.413756366
#> 77   26    2 -0.48291103  1.00404451 -0.236662345
#> 78   26    3  0.98477614  0.40857769  1.423027822
#> 79   27    1  0.91543129 -0.46764781  0.296901388
#> 81   27    3 -0.22780362  0.33442452  1.128644605
#> 82   28    1  0.69582402 -0.85803491 -0.689318195
#> 84   28    3  0.68273404  0.81650754 -0.073601914
#> 86   29    2 -1.25645902  0.71420802 -0.714352800
#> 88   30    1 -0.73486556  0.98939456 -0.252660832
#> 90   30    3  0.41641349 -1.26885749  0.043178765
#> 91   31    1  0.89197464  1.32201695 -0.400113902
#> 92   31    2 -0.36235234 -0.56387952  1.605884870
#> 93   31    3 -0.14885530  1.76166101  0.997222081
#> 94   32    1  0.47356885 -0.64997134 -0.112449170
#> 95   32    2  1.05168882  0.63878506 -0.278007889
#> 96   32    3  0.87068072  0.14152028 -0.180395976
#> 97   33    1 -1.36062567 -0.07129532  1.257724412
#> 100  34    1  0.26665271  1.46749574 -0.957902921
#> 102  34    3  0.54418631 -0.49551504 -0.077556987
#> 103  35    1 -0.73160010  0.24849737 -0.823522734
#> 104  35    2 -0.07715187 -1.02757136 -2.246226772
#> 105  35    3  2.01073460 -0.04396826  0.771145589
#> 107  36    2 -1.14374243 -0.07647743  1.275909695
#> 108  36    3 -1.58326599 -0.35995922 -1.778436264
#> 109  37    1  0.73741306  0.19604791 -1.362837841
#> 110  37    2  0.95496015  0.99691125 -0.678983392
#> 111  37    3 -0.76147549  0.36628567 -1.280547203
#> 112  38    1 -1.80175826  0.69089766 -1.569678726
#> 113  38    2 -1.38475978 -0.87613525 -1.901778298
#> 114  38    3 -0.29768182 -0.50910335  0.874623154
#> 115  39    1  0.73420236 -0.62946316  0.034898764
#> 117  39    3 -0.55886357 -1.99258928 -1.357406375
#> 118  40    1 -0.27746832  0.48024268  1.346344914
#> 119  40    2  0.59006113  0.18226741 -0.566846144
#> 120  40    3 -0.02531491  1.46305754 -0.313826957
#> 121  41    1 -1.20992250 -1.50110155 -0.199340904
#> 122  41    2 -0.07770139  1.11979077 -1.400625825
#> 123  41    3 -0.60569915 -1.07878348 -1.231706778
#> 124  42    1 -0.43986955  0.33874400 -0.338890531
#> 125  42    2  0.54086764 -0.15415391  0.341898974
#> 126  42    3 -0.89992214 -1.06370457 -1.022483842
#> 127  43    1 -1.56684238 -0.19415249  0.104519257
#> 129  43    3 -0.38155242  1.65133257 -1.131834251
#> 130  44    1  1.89755773  1.04585205 -0.792372374
#> 133  45    1 -0.16731511 -0.69969700  0.958550224
#> 135  45    3  0.06014704  2.03110961  0.707414736
#> 136  46    1  0.26975570  0.67371806 -1.404740854
#> 138  46    3  0.37878777 -0.01397260 -0.446644356
#> 140  47    2  0.66631976 -0.80748218  1.250762230
#> 141  47    3  0.43339214 -0.55598739  0.517442156
#> 143  48    2 -1.37233726 -2.01952392 -0.610677150
#> 144  48    3 -0.98356077 -0.21489672  1.635037280
#> 146  49    2 -0.93981976 -0.81192753  0.841265366
#> 148  50    1  1.22519259 -1.07744912  0.789071705
#> 149  50    2 -1.04938406  1.01421809  0.392156926
#> 150  50    3 -0.11281556  0.53169112  0.442879772
#> 151  51    1 -0.32714874  0.74188351  0.650292804
#> 153  51    3 -0.44646703 -0.14496339  1.004326647
#> 156  52    3  0.15650419  0.03121845  0.601126945
#> 157  53    1 -1.15758334 -0.48682666 -0.124373045
#> 158  53    2  0.57212722 -0.26056529  1.089070496
#> 159  53    3 -0.61635461  1.15150374  0.170800278
#> 160  54    1  0.31623970 -0.71795193 -0.580530337
#> 162  54    3  0.16947352 -0.58757878 -2.007412536
#> 163  55    1 -0.47583758 -0.75841782 -1.234208277
#> 164  55    2 -0.66680543  1.25606537 -0.348280844
#> 167  56    2  0.54319453 -0.28969276  0.940967479
#> 170  57    2  0.68735178  0.24384304 -0.522096078
#> 171  57    3 -0.96460769  0.05719409  0.308478421
#> 172  58    1  0.22744428  0.39119382 -0.528083292
#> 173  58    2  1.84971532  1.18829834  0.403735855
#> 175  59    1  1.33453976  0.19514458 -0.042580160
#> 177  59    3 -0.36794402  0.59570991 -3.006342876
#> 178  60    1 -0.12367872 -0.08125807  0.780491543
#> 179  60    2 -0.34417383  0.88126900  0.245144534
#> 181  61    1  1.32197390  2.37091477  0.672508933
#> 184  62    1 -0.40256406 -0.43256653  0.217144785
#> 185  62    2  0.79179127 -0.67300936  1.613662196
#> 187  63    1 -0.77408936 -0.39675974  2.574740159
#> 188  63    2 -2.67452586  0.14676967 -0.516565635
#> 190  64    1 -0.49524216 -0.56528319 -1.118766755
#> 191  64    2  0.49661664  0.17007117  1.837863360
#> 192  64    3  1.21732856 -0.29117703  0.641805441
#> 196  66    1 -0.42276263  0.56017442 -0.323582682
#> 197  66    2  0.13444519  0.85967083  0.388928705
#> 198  66    3  2.04440812 -0.84764967  1.105069027
#> 199  67    1 -0.75852344  2.02262308 -0.021861849
#> 200  67    2 -1.04710637 -0.02478192 -0.603689778
#> 201  67    3  0.29094018  0.04273221 -1.176843191
#> 203  68    2 -0.52028192 -0.64165838  0.971764440
#> 204  68    3  1.69667178 -0.24059417 -0.035337758
#> 205  69    1  2.04796436  0.01744152  0.490847715
#> 206  69    2  1.98098662  0.18482853  1.235264662
#> 207  69    3 -0.50981686  0.85458112  0.035466423
#> 209  70    2  0.99546597  0.56724777  0.222053638
#> 211  71    1  0.44564626  0.27441875  2.396770301
#> 212  71    2 -1.37384585  0.11607404 -0.927745456
#> 215  72    2 -0.15387745  2.08954243 -1.487650977
#> 216  72    3  1.56696424  2.52713527  1.327347626
#> 220  74    1 -0.31059035  0.01684000 -0.299589591
#> 221  74    2 -0.15146275  0.37528387  0.017211530
#> 222  74    3  0.08898270  1.08248947 -0.672273397
#> 223  75    1 -1.45037365  1.18400856 -0.260355391
#> 224  75    2  0.84298116 -1.88052053  0.113998687
#> 225  75    3  1.36025230  1.72953938  0.324486792
#> 227  76    2  1.30346859  1.03076352  0.844244201
#> 228  76    3  0.33291375 -0.37251595  0.710611432
#> 230  77    2  0.89534412 -1.19113241  1.561727361
#> 234  78    3 -0.11658391 -0.32166439 -0.313911442
#> 235  79    1 -0.31220221  0.42452723 -1.074727061
#> 237  79    3 -0.31138401  1.65183408  0.244615940
#> 238  80    1 -0.59945530  0.47469849 -1.251611331
#> 240  80    3  0.91241433  1.25971301  1.278662106
#> 241  81    1  0.09778511  0.39730883 -1.094054572
#> 242  81    2 -0.53678456 -0.30969780 -0.455975345
#> 243  81    3  0.10178658  0.82591536  0.290276155
#> 244  82    1  0.14769492  0.30449601  0.223043436
#> 245  82    2 -1.33520701 -1.40896159  2.320008018
#> 246  82    3  1.29192650 -0.32305666 -0.774740691
#> 247  83    1  2.01723437  1.10475613  2.808067922
#> 248  83    2 -0.36189542 -0.77370097 -0.755960624
#> 251  84    2  0.23496616 -1.00618307  0.983034981
#> 252  84    3 -0.42953123 -0.69546606  1.184164414
#> 253  85    1 -0.07216856 -0.48620686  0.994891650
#> 255  85    3  0.65669226  1.07821141 -0.229395923
#> 258  86    3 -0.03701812  1.19133929 -0.595170773
#> 261  87    3 -1.47167575  1.36533384 -0.310062141
#> 263  88    2 -0.23075934  0.72462563  1.027384325
#> 264  88    3  0.13432327  0.70793034 -1.489916488
#> 265  89    1 -0.50702660  0.98330637  0.097752902
#> 267  89    3  1.12277235  0.54674412 -0.821170042
#> 273  91    3  0.22731656 -0.35024803  0.045719649
#> 275  92    2  0.53433165 -0.73970114 -0.329350529
#> 276  92    3  0.91915273 -1.52743114  0.168828394
#> 278  93    2 -0.97961270  0.64102438 -0.998421828
#> 279  93    3 -0.95210323  0.96187752  0.194344736
#> 280  94    1 -0.16323012 -0.44926268 -0.427944031
#> 281  94    2 -0.39997239 -2.28607900 -0.111651217
#> 287  96    2  0.65410422 -0.96758472 -0.468199511
#> 288  96    3 -0.31564225 -0.45212830  0.036386642
#> 289  97    1  1.00088061  0.45374545 -0.298738168
#> 293  98    2 -0.82557185  0.37548887 -0.094946186
#> 295  99    1  0.24881942  0.67559274  0.329577842
#> 296  99    2 -0.64709300  0.91476176  0.102973499
#> 297  99    3  0.41816741 -0.41675436 -0.748540920
#> 298 100    1  0.56381769  0.04251611 -0.035888525
#> 299 100    2 -0.86967011  1.79410517 -1.369178436