Drop NA rows.

Drop rows with all missing (NA) values.

Usage

dropRowsWithAllNA(data, ignore = NULL)

Arguments

data

Dataframe to drop rows from.

ignore

Names of columns to ignore for determining whether each row had all missing values.

Value

A dataframe with rows removed that had all missing values in non-ignored columns.

Details

Drop rows that have no observed values, i.e., all values in the row are missing (NA), excluding the ignored columns.

See also

Examples

# Prepare Data
df <- expand.grid(ID = 1:100, time = c(1, 2, 3))
df <- df[order(df$ID),]
row.names(df) <- NULL
df$score1 <- rnorm(nrow(df))
df$score2 <- rnorm(nrow(df))
df$score3 <- rnorm(nrow(df))
df[sample(1:nrow(df), size = 100), c("score1","score2","score3")] <- NA

# Drop Rows with All NA in Non-Ignored Columns
dropRowsWithAllNA(df, ignore = c("ID","time"))
#>      ID time       score1      score2       score3
#> 1     1    1  0.689146390 -0.28963790  1.411870873
#> 5     2    2 -0.429310665  0.37381260  0.362078075
#> 6     2    3  0.273539055  0.48360543  0.700960895
#> 8     3    2  1.668139992  0.99245161  1.139158270
#> 10    4    1 -0.563607063  0.63374342 -0.007159302
#> 11    4    2 -0.048200297  0.30130406 -0.989717038
#> 12    4    3 -0.644849008  0.53540384 -0.214636639
#> 14    5    2  1.256314559  1.95536073  0.013072450
#> 15    5    3 -1.222501732 -0.35029394 -1.006662688
#> 16    6    1 -0.306347069  1.43845252 -0.444209491
#> 18    6    3  0.087155047 -1.02519855  0.638901159
#> 19    7    1 -1.106746484 -0.04390200  0.580394014
#> 20    7    2  0.509824764 -0.32434552 -1.245153828
#> 21    7    3  1.189780884  0.31934744 -1.191274936
#> 25    9    1 -0.301428596 -0.99308722  0.052501354
#> 26    9    2 -1.549479609  0.72410853  1.200836500
#> 28   10    1  0.210963791 -0.59416315 -0.551985787
#> 29   10    2  0.479902641 -0.56590379  0.941191015
#> 33   11    3  1.562969473  0.50153864 -0.043803878
#> 34   12    1 -1.561180624 -0.06052325  0.271921865
#> 35   12    2 -1.006295189  0.17757223  2.497815829
#> 38   13    2  0.519088294 -1.60657709  0.406992525
#> 40   14    1 -0.561724812 -0.95783909 -0.332577014
#> 41   14    2  0.506200163  0.15861449  0.265610157
#> 42   14    3 -0.470121057  1.81618883  0.163216140
#> 43   15    1  1.052720658  0.42402218 -0.934675712
#> 44   15    2 -0.548255611  0.75353450  0.518930279
#> 45   15    3 -1.815716631  1.28912411 -0.158326401
#> 46   16    1 -1.025024490  1.02151744  0.535134099
#> 47   16    2 -1.505419557  0.34032502 -1.286095649
#> 48   16    3  1.148480165  0.10249559  1.039636333
#> 49   17    1 -0.235446281 -0.92318861 -0.220851435
#> 50   17    2  0.980511471  1.84860068  0.707972016
#> 51   17    3 -0.681326701  0.17954935 -0.072282608
#> 52   18    1 -0.535285705  0.80626799 -0.196291321
#> 53   18    2  0.559305399  2.18546792 -0.271844917
#> 54   18    3  1.725757591 -1.75716887  0.793323375
#> 57   19    3  1.350713531 -0.22886523 -0.455143131
#> 58   20    1 -0.806236652 -1.14973765  1.463705935
#> 59   20    2 -0.414731855  0.26695992 -2.355188157
#> 60   20    3  0.043142643  0.65822671 -1.130516761
#> 63   21    3 -0.066731130 -1.04967722  1.219572694
#> 65   22    2 -0.406599965  1.40121241  0.439299875
#> 66   22    3 -0.685567218  0.52342579 -0.311222883
#> 67   23    1  0.956675837 -0.31820181 -0.199988863
#> 68   23    2 -0.226919831  1.12657695  1.662453852
#> 69   23    3 -1.121861803 -0.36408292 -0.348245415
#> 70   24    1  0.417256996 -0.78366728  0.866831654
#> 71   24    2  0.905750164 -0.89026968  1.094056230
#> 72   24    3  0.291960480 -1.06471645  0.989521227
#> 74   25    2 -0.629245341 -0.66755520 -0.413756366
#> 75   25    3 -0.482911031  1.00404451 -0.236662345
#> 76   26    1  0.984776138  0.40857769  1.423027822
#> 77   26    2  0.915431286 -0.46764781  0.296901388
#> 78   26    3 -0.891199540  1.22074106 -0.156800417
#> 79   27    1 -0.227803616  0.33442452  1.128644605
#> 81   27    3 -0.323946977  0.08051104 -0.049112903
#> 82   28    1  0.682734036  0.81650754 -0.073601914
#> 84   28    3 -1.256459021  0.71420802 -0.714352800
#> 86   29    2 -0.734865560  0.98939456 -0.252660832
#> 88   30    1  0.416413489 -1.26885749  0.043178765
#> 90   30    3 -0.362352344 -0.56387952  1.605884870
#> 91   31    1 -0.148855303  1.76166101  0.997222081
#> 92   31    2  0.473568852 -0.64997134 -0.112449170
#> 93   31    3  1.051688823  0.63878506 -0.278007889
#> 94   32    1  0.870680724  0.14152028 -0.180395976
#> 95   32    2 -1.360625665 -0.07129532  1.257724412
#> 96   32    3  0.561632000  0.82937151 -0.555409584
#> 97   33    1 -0.057109572  0.18699052 -0.920261813
#> 100  34    1  0.544186309 -0.49551504 -0.077556987
#> 102  34    3 -0.077151868 -1.02757136 -2.246226772
#> 103  35    1  2.010734604 -0.04396826  0.771145589
#> 104  35    2  1.157662754 -0.92988060 -0.586150780
#> 105  35    3 -1.143742433 -0.07647743  1.275909695
#> 107  36    2  0.737413064  0.19604791 -1.362837841
#> 108  36    3  0.954960154  0.99691125 -0.678983392
#> 109  37    1 -0.761475492  0.36628567 -1.280547203
#> 110  37    2 -1.801758259  0.69089766 -1.569678726
#> 111  37    3 -1.384759781 -0.87613525 -1.901778298
#> 112  38    1 -0.297681818 -0.50910335  0.874623154
#> 113  38    2  0.734202361 -0.62946316  0.034898764
#> 114  38    3 -0.361433208 -0.05134831  0.332989786
#> 117  39    3  0.590061125  0.18226741 -0.566846144
#> 118  40    1 -0.025314909  1.46305754 -0.313826957
#> 119  40    2 -1.209922502 -1.50110155 -0.199340904
#> 120  40    3 -0.077701387  1.11979077 -1.400625825
#> 121  41    1 -0.605699153 -1.07878348 -1.231706778
#> 122  41    2 -0.439869547  0.33874400 -0.338890531
#> 123  41    3  0.540867642 -0.15415391  0.341898974
#> 124  42    1 -0.899922143 -1.06370457 -1.022483842
#> 125  42    2 -1.566842379 -0.19415249  0.104519257
#> 126  42    3 -0.001528403 -0.49651922 -0.112296543
#> 127  43    1 -0.381552421  1.65133257 -1.131834251
#> 129  43    3  0.398168313  0.57291928 -0.265746858
#> 130  44    1  0.679407341 -1.57845898  0.667765125
#> 133  45    1  0.060147043  2.03110961  0.707414736
#> 135  45    3 -0.825699295  1.26665727  0.794059229
#> 136  46    1  0.378787773 -0.01397260 -0.446644356
#> 138  46    3  0.666319762 -0.80748218  1.250762230
#> 140  47    2  0.914591396 -0.46384001 -0.877240027
#> 141  47    3 -1.372337261 -2.01952392 -0.610677150
#> 143  48    2 -0.794677313 -1.32004024 -0.945749477
#> 144  48    3 -0.939819756 -0.81192753  0.841265366
#> 146  49    2  1.225192587 -1.07744912  0.789071705
#> 148  50    1 -0.112815561  0.53169112  0.442879772
#> 149  50    2 -0.327148739  0.74188351  0.650292804
#> 150  50    3  0.511212794 -0.16025589  0.248377533
#> 151  51    1 -0.446467027 -0.14496339  1.004326647
#> 153  51    3  0.444805504 -1.26750352 -0.353720918
#> 156  52    3  0.572127218 -0.26056529  1.089070496
#> 157  53    1 -0.616354607  1.15150374  0.170800278
#> 158  53    2  0.316239702 -0.71795193 -0.580530337
#> 159  53    3 -0.094991681 -0.80949978  0.093760686
#> 160  54    1  0.169473519 -0.58757878 -2.007412536
#> 161  54    2 -0.475837583 -0.75841782 -1.234208277
#> 162  54    3 -0.666805426  1.25606537 -0.348280844
#> 163  55    1  1.056044324 -0.45300889 -1.081510052
#> 164  55    2  0.384280892 -0.64639512  0.421888999
#> 167  56    2 -1.191244185 -1.11170959  1.110108336
#> 170  57    2  0.227444275  0.39119382 -0.528083292
#> 171  57    3  1.849715316  1.18829834  0.403735855
#> 172  58    1 -0.642118740  0.41710581  1.373026156
#> 173  58    2  1.334539758  0.19514458 -0.042580160
#> 175  59    1 -0.367944016  0.59570991 -3.006342876
#> 177  59    3 -0.344173827  0.88126900  0.245144534
#> 178  60    1  0.728273311  0.94490127  0.990826830
#> 179  60    2  1.321973899  2.37091477  0.672508933
#> 181  61    1  0.474416948  0.45842063  2.097213395
#> 184  62    1 -0.963065337 -0.68249662 -2.548375111
#> 185  62    2 -0.774089357 -0.39675974  2.574740159
#> 187  63    1  0.167525707  0.66893248 -0.619008194
#> 188  63    2 -0.495242157 -0.56528319 -1.118766755
#> 190  64    1  1.217328555 -0.29117703  0.641805441
#> 191  64    2 -0.536413771  1.86553257 -0.310903840
#> 192  64    3  0.510848723 -0.50717361  0.014340688
#> 196  66    1  2.044408123 -0.84764967  1.105069027
#> 197  66    2 -0.758523443  2.02262308 -0.021861849
#> 198  66    3 -1.047106372 -0.02478192 -0.603689778
#> 199  67    1  0.290940178  0.04273221 -1.176843191
#> 200  67    2 -1.523579363  0.56328774  0.418225014
#> 201  67    3 -0.520281917 -0.64165838  0.971764440
#> 203  68    2  2.047964363  0.01744152  0.490847715
#> 204  68    3  1.980986617  0.18482853  1.235264662
#> 205  69    1 -0.509816865  0.85458112  0.035466423
#> 206  69    2 -1.054151581 -2.14941880  0.319541298
#> 207  69    3  0.995465973  0.56724777  0.222053638
#> 209  70    2  0.445646258  0.27441875  2.396770301
#> 211  71    1  0.733605707 -0.77136270  0.557204266
#> 212  71    2  0.572860596  1.72291963  0.643652284
#> 214  72    1  1.566964243  2.52713527  1.327347626
#> 215  72    2 -0.058145503  0.54316887 -0.713110417
#> 216  72    3  1.490530036  1.44959051 -1.184780242
#> 220  74    1  0.088982698  1.08248947 -0.672273397
#> 221  74    2 -1.450373646  1.18400856 -0.260355391
#> 222  74    3  0.842981156 -1.88052053  0.113998687
#> 223  75    1  1.360252302  1.72953938  0.324486792
#> 225  75    3  1.303468590  1.03076352  0.844244201
#> 228  76    3  0.895344115 -1.19113241  1.561727361
#> 230  77    2  0.803073894 -0.96608851  0.833568276
#> 234  78    3 -0.086431889  0.43928272 -0.302506944
#> 235  79    1 -0.311384010  1.65183408  0.244615940
#> 237  79    3  0.843275041  0.09513949 -1.733921415
#> 238  80    1  0.912414333  1.25971301  1.278662106
#> 240  80    3 -0.536784560 -0.30969780 -0.455975345
#> 241  81    1  0.101786578  0.82591536  0.290276155
#> 242  81    2  0.147694919  0.30449601  0.223043436
#> 243  81    3 -1.335207011 -1.40896159  2.320008018
#> 244  82    1  1.291926497 -0.32305666 -0.774740691
#> 245  82    2  2.017234373  1.10475613  2.808067922
#> 246  82    3 -0.361895417 -0.77370097 -0.755960624
#> 247  83    1 -0.143103897  0.89568062  0.229211605
#> 248  83    2 -0.711301049  0.98024952  0.097527484
#> 250  84    1 -0.429531231 -0.69546606  1.184164414
#> 251  84    2 -0.072168560 -0.48620686  0.994891650
#> 252  84    3  0.015891336 -0.35724726 -0.417221145
#> 253  85    1  0.656692261  1.07821141 -0.229395923
#> 256  86    1 -0.037018120  1.19133929 -0.595170773
#> 257  86    2  0.186111684  0.94689258  1.498313908
#> 258  86    3 -2.124311597  1.17958131 -0.634377790
#> 260  87    2 -0.409724927  0.54395009  1.415969444
#> 261  87    3 -0.230759342  0.72462563  1.027384325
#> 263  88    2 -0.507026601  0.98330637  0.097752902
#> 264  88    3  2.023726295  0.60664997 -0.274708562
#> 265  89    1  1.122772350  0.54674412 -0.821170042
#> 266  89    2  0.946122654  0.66984579 -0.988737886
#> 267  89    3 -2.135465869 -0.44353073  1.135918365
#> 271  91    1  0.227316561 -0.35024803  0.045719649
#> 275  92    2 -0.325150482  0.46537184 -0.434508733
#> 276  92    3 -0.979612701  0.64102438 -0.998421828
#> 280  94    1  0.192011842 -0.79424640 -1.980097643
#> 281  94    2 -0.281151314 -1.73710520  0.009597104
#> 283  95    1 -1.034878996  0.34278914 -0.296573292
#> 289  97    1 -0.015952454 -1.04146183 -0.690588407
#> 291  97    3 -0.825571851  0.37548887 -0.094946186
#> 292  98    1 -0.961128624  0.35845352 -1.882785636
#> 293  98    2  0.248819418  0.67559274  0.329577842
#> 294  98    3 -0.647092998  0.91476176  0.102973499
#> 297  99    3 -0.869670110  1.79410517 -1.369178436
#> 298 100    1 -0.562077300  0.47395801  0.212756151
#> 299 100    2  1.234077644 -0.55123963 -0.887887107